Crate re_viewer

Rerun Viewer GUI.

This crate contains all the GUI code for the Rerun Viewer, including all 2D and 3D visualization code.

Modules

app 🔒

app_icon 🔒

env_vars

external

math

misc 🔒

native 🔒

remote_viewer_app 🔒

stream_rrd_from_http

ui 🔒

viewer_analytics 🔒

All telemetry analytics collected by the Rerun Viewer are defined in this file for easy auditing.

Structs

App

The Rerun viewer as an [eframe] application.

Profiler

Wraps a connection to a [puffin] viewer.

RemoteViewerApp

Connects to a server over WebSockets.

StartupOptions

Settings set once at startup (e.g. via command-line options) and not serialized.

Enums

AppEnvironment

Where is this App running in?

Constants

APPLICATION_NAME 🔒

Functions

customize_eframe 🔒

hardware_tier 🔒

run_native_app

run_native_viewer_with_messages

wake_up_ui_thread_on_each_msg

This wakes up the ui thread each time we receive a new message.

wgpu_options 🔒