Module re_viewer::app

Structs

App

The Rerun viewer as an [eframe] application.

AppState ğŸ”’

StartupOptions

Settings set once at startup (e.g. via command-line options) and not serialized.

Enums

PanelSelection ğŸ”’

TimeControlCommand ğŸ”’

Constants

FILE_SAVER_PROMISE ğŸ”’

MAX_ZOOM_FACTOR ğŸ”’

MIN_ZOOM_FACTOR ğŸ”’

WATERMARK ğŸ”’

Functions

about_rerun_ui ğŸ”’

debug_menu_options_ui ğŸ”’

egui_debug_options_ui ğŸ”’

file_saver_progress_ui ğŸ”’

frame_time_label_ui ğŸ”’

input_latency_label_ui ğŸ”’

load_file_contents ğŸ”’

load_file_path ğŸ”’

load_rrd_to_log_db ğŸ”’

main_view_selector_ui ğŸ”’

memory_use_label_ui ğŸ”’

new_recording_confg ğŸ”’

open ğŸ”’

options_menu_ui ğŸ”’

paint_background_fill ğŸ”’

paint_native_window_frame ğŸ”’

preview_files_being_dropped ğŸ”’

recording_config_entry ğŸ”’

recordings_menu ğŸ”’

rerun_menu_button_ui ğŸ”’

save ğŸ”’

save_buttons_ui ğŸ”’

save_database_to_file ğŸ”’

Returns a closure that, when run, will save the contents of the current database to disk, at the specified path.

top_bar_ui ğŸ”’

top_panel ğŸ”’

wait_screen_ui ğŸ”’

warning_panel ğŸ”’