re_viewer
Module blueprint

Auto-generated blueprint-related types.

They all implement the re_types_core::Component trait.

Unstable. Used for the ongoing blueprint experimentations.

Modules

validation

validation_gen 🔒

Functions

is_valid_blueprint

Because blueprints are both read and written the schema must match what we expect to find or else we will run into all kinds of problems.