re_viewer::blueprint
Module validation_gen

Structs

ActiveTab

Component: The active tab in a tabbed container.

ApplyLatestAt

Component: Whether empty cells in a dataframe should be filled with a latest-at query.

AutoLayout

Component: Whether the viewport layout is determined automatically.

AutoViews

Component: Whether or not views should be created automatically.

ColumnShare

Component: The layout share of a column in the container.

ComponentColumnSelector

Component: Describe a component column to be selected in the dataframe view.

Enabled

Component: Whether a procedure is enabled.

FilterByRange

Component: Configuration for a filter-by-range feature of the dataframe view.

FilterIsNotNull

Component: Configuration for the filter is not null feature of the dataframe view.

ForceDistance

Component: The target distance between two nodes.

ForceIterations

Component: Specifies how often this force should be applied per iteration.

ForceStrength

Component: The strength of a given force.

GridColumns

Component: How many columns a grid container should have.

GridSpacing

Component: Space between grid lines of one line to the next in scene units.

IncludedContent

Component: All the contents in the container.

LockRangeDuringZoom

Component: Indicate whether the range should be locked when zooming in on the data.

NearClipPlane

Component: Distance to the near clip plane used for Spatial2DView.

QueryExpression

Component: An individual query expression used to filter a set of datatypes::EntityPaths.

RootContainer

Component: The container that sits at the root of a viewport.

RowShare

Component: The layout share of a row in the container.

SelectedColumns

Component: Describe a component column to be selected in the dataframe view.

TensorDimensionIndexSlider

Component: Show a slider for the index of some dimension of a slider.

TimelineName

Component: A timeline identified by its name.

ViewClass

Component: The class identifier of view, e.g. "2D", "TextLog", ….

ViewMaximized

Component: Whether a view is maximized.

ViewOrigin

Component: The origin of a view.

ViewerRecommendationHash

Component: Hash of a viewer recommendation.

VisibleTimeRange

Component: The range of values on a given timeline that will be included in a view’s query.

VisualBounds2D

Component: Visual bounds in 2D space used for Spatial2DView.

VisualizerOverride

Component: Single visualizer override the visualizers for an entity.

ZoomLevel

Component: A zoom level determines how much of the world is visible on a map.

Enums

BackgroundKind

Component: The type of the background in a view.

ContainerKind

Component: The kind of a blueprint container (tabs, grid, …).

Corner2D

Component: One of four 2D corners, typically used to align objects.

MapProvider

Component: Name of the map provider to be used in Map views.

PanelState

Component: Tri-state for panel controls.

ViewFit

Component: Determines whether an image or texture should be scaled to fit the viewport.

Functions

is_valid_blueprint

Because blueprints are both read and written the schema must match what we expect to find or else we will run into all kinds of problems.