re_viewer::event
Module serde

Customs serialization for event payloads.

We serialize events into JSON when crossing the js/py language bridge, and some types don’t serialize into something that would be useful “as-is” in those languages.

Modules

blueprint_id

entity_path

instance_id

recording_id

time_real

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Deserializer

A data format that can deserialize any data structure supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Serializer

A data format that can serialize any data structure supported by Serde.

Derive Macros

Deserialize

Serialize