Module re_viewer::external

Re-exports

pub use eframe;

pub use egui;

pub use re_chunk;

pub use re_chunk_store;

pub use re_data_ui;

pub use re_entity_db;

pub use re_log;

pub use re_log_types;

pub use re_memory;

pub use re_renderer;

pub use re_types;

pub use re_ui;

pub use re_viewer_context;

pub use re_viewport;

Modules

arrow2

Welcome to arrow2’s documentation. Thanks for checking it out!

crossbeam

Tools for concurrent programming.

nohash_hasher

re_chunk

A chunk of Rerun data, encoded using Arrow. Used for logging, transport, storage and compute.

re_chunk_store

The Rerun chunk store, implemented on top of Apache Arrow using the arrow2 crate.

re_entity_db

This is how we store and index logging data.

re_log_encoding

Crate that handles encoding of rerun log types.

re_log_types

The different types that make up the rerun log format.

re_space_view

Rerun Space View utilities

re_types_blueprint

The core types and traits that power Rerun’s Blueprint sub-system.

re_ui

Rerun GUI theme and helpers, built around egui.