re_viewer::external::arrow::compute
Function filter_record_batch

pub fn filter_record_batch(
    record_batch: &RecordBatch,
    predicate: &BooleanArray,
) -> Result<RecordBatch, ArrowError>

Returns a filtered RecordBatch where the corresponding elements of predicate are true.

This is the equivalent of calling filter on each column of the RecordBatch.