re_viewer::external::arrow::ipc::gen::Message
Function size_prefixed_root_as_message

pub fn size_prefixed_root_as_message(
    buf: &[u8],
) -> Result<Message<'_>, InvalidFlatbuffer>

Verifies that a buffer of bytes contains a size prefixed Message and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_message_unchecked.