Struct re_viewer::external::arrow2::array::BinaryArray

pub struct BinaryArray<O>where
    O: Offset,{
    data_type: DataType,
    offsets: OffsetsBuffer<O>,
    values: Buffer<u8>,
    validity: Option<Bitmap>,
}

A BinaryArray is Arrow’s semantically equivalent of an immutable Vec<Option<Vec<u8>>>. It implements Array.

The size of this struct is O(1), as all data is stored behind an std::sync::Arc.

Example

use arrow2::array::BinaryArray;
use arrow2::bitmap::Bitmap;
use arrow2::buffer::Buffer;

let array = BinaryArray::<i32>::from([Some([1, 2].as_ref()), None, Some([3].as_ref())]);
assert_eq!(array.value(0), &[1, 2]);
assert_eq!(array.iter().collect::<Vec<_>>(), vec![Some([1, 2].as_ref()), None, Some([3].as_ref())]);
assert_eq!(array.values_iter().collect::<Vec<_>>(), vec![[1, 2].as_ref(), &[], &[3]]);
// the underlying representation:
assert_eq!(array.values(), &Buffer::from(vec![1, 2, 3]));
assert_eq!(array.offsets().buffer(), &Buffer::from(vec![0, 2, 2, 3]));
assert_eq!(array.validity(), Some(&Bitmap::from([true, false, true])));

Generic parameter

The generic parameter Offset can only be i32 or i64 and tradeoffs maximum array length with memory usage:

the sum of lengths of all elements cannot exceed Offset::MAX

the total size of the underlying data is array.len() * size_of::<Offset>() + sum of lengths of all elements

Safety

The following invariants hold:

Two consecutives offsets casted (as) to usize are valid slices of values.

len is equal to validity.len(), when defined.

Fields
data_type: DataTypeoffsets: OffsetsBuffer<O>values: Buffer<u8>validity: Option<Bitmap>
Implementations

impl<O> BinaryArray<O>where O: Offset,

pub fn try_new( data_type: DataType, offsets: OffsetsBuffer<O>, values: Buffer<u8>, validity: Option<Bitmap> ) -> Result<BinaryArray<O>, Error>

Returns a BinaryArray created from its internal representation.

Errors

This function returns an error iff:

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len().

The data_type’s crate::datatypes::PhysicalType is not equal to either Binary or LargeBinary.

Implementation

This function is O(1)

pub fn from_slice<T, P>(slice: P) -> BinaryArray<O>where T: AsRef<[u8]>, P: AsRef<[T]>,

Creates a new BinaryArray from slices of &[u8].

pub fn from<T, P>(slice: P) -> BinaryArray<O>where T: AsRef<[u8]>, P: AsRef<[Option<T>]>,

Creates a new BinaryArray from a slice of optional &[u8].

pub fn iter( &self ) -> ZipValidity<&[u8], ArrayValuesIter<'_, BinaryArray<O>>, BitmapIter<'_>>

Returns an iterator of Option<&[u8]> over every element of this array.

pub fn values_iter(&self) -> ArrayValuesIter<'_, BinaryArray<O>>

Returns an iterator of &[u8] over every element of this array, ignoring the validity

pub fn len(&self) -> usize

Returns the length of this array

pub fn value(&self, i: usize) -> &[u8]

Returns the element at index i

Panics

iff i >= self.len()

pub unsafe fn value_unchecked(&self, i: usize) -> &[u8]

Returns the element at index i

Safety

Assumes that the i < self.len.

pub fn get(&self, i: usize) -> Option<&[u8]>

Returns the element at index i or None if it is null

Panics

iff i >= self.len()

pub fn data_type(&self) -> &DataType

Returns the DataType of this array.

pub fn values(&self) -> &Buffer<u8>

Returns the values of this BinaryArray.

pub fn offsets(&self) -> &OffsetsBuffer<O>

Returns the offsets of this BinaryArray.

pub fn validity(&self) -> Option<&Bitmap>

The optional validity.

pub fn slice(&mut self, offset: usize, length: usize)

Slices this BinaryArray.

Implementation

This function is O(1).

Panics

iff offset + length > self.len().

pub unsafe fn slice_unchecked(&mut self, offset: usize, length: usize)

Slices this BinaryArray.

Implementation

This function is O(1).

Safety

The caller must ensure that offset + length <= self.len().

pub fn sliced(self, offset: usize, length: usize) -> BinaryArray<O>

Returns this array sliced.

Implementation

This function is O(1).

Panics

iff offset + length > self.len().

pub fn sliced_unchecked(self, offset: usize, length: usize) -> BinaryArray<O>

Returns this array sliced.

Implementation

This function is O(1).

Safety

The caller must ensure that offset + length <= self.len().

pub fn with_validity(self, validity: Option<Bitmap>) -> BinaryArray<O>

Returns this array with a new validity.

Panic

Panics iff validity.len() != self.len().

pub fn set_validity(&mut self, validity: Option<Bitmap>)

Sets the validity of this array.

Panics

This function panics iff values.len() != self.len().

pub fn boxed(self) -> Box<dyn Array, Global>

Boxes this array into a Box<dyn Array>.

pub fn arced(self) -> Arc<dyn Array>

Arcs this array into a std::sync::Arc<dyn Array>.

pub fn into_mut(self) -> Either<BinaryArray<O>, MutableBinaryArray<O>>

Try to convert this BinaryArray to a MutableBinaryArray

pub fn new_empty(data_type: DataType) -> BinaryArray<O>

Creates an empty BinaryArray, i.e. whose .len is zero.

pub fn new_null(data_type: DataType, length: usize) -> BinaryArray<O>

Creates an null BinaryArray, i.e. whose .null_count() == .len().

pub fn default_data_type() -> DataType

Returns the default DataType, DataType::Binary or DataType::LargeBinary

pub fn new( data_type: DataType, offsets: OffsetsBuffer<O>, values: Buffer<u8>, validity: Option<Bitmap> ) -> BinaryArray<O>

Alias for unwrapping Self::try_new

pub fn from_trusted_len_values_iter<T, I>(iterator: I) -> BinaryArray<O>where T: AsRef<[u8]>, I: TrustedLen<Item = T>,

Returns a BinaryArray from an iterator of trusted length.

The BinaryArray is guaranteed to not have a validity

pub fn from_iter_values<T, I>(iterator: I) -> BinaryArray<O>where T: AsRef<[u8]>, I: Iterator<Item = T>,

Returns a new BinaryArray from a Iterator of &[u8].

The BinaryArray is guaranteed to not have a validity

pub unsafe fn from_trusted_len_iter_unchecked<I, P>( iterator: I ) -> BinaryArray<O>where P: AsRef<[u8]>, I: Iterator<Item = Option<P>>,

Creates a BinaryArray from an iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> BinaryArray<O>where P: AsRef<[u8]>, I: TrustedLen<Item = Option<P>>,

Creates a BinaryArray from a TrustedLen

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>( iterator: I ) -> Result<BinaryArray<O>, E>where P: AsRef<[u8]>, I: IntoIterator<Item = Result<Option<P>, E>>,

Creates a BinaryArray from an falible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>(iter: I) -> Result<BinaryArray<O>, E>where P: AsRef<[u8]>, I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a BinaryArray from an fallible iterator of trusted length.