Module re_viewer::external::arrow2::compute::arity

Defines kernels suitable to perform operations to primitive arrays.

Functions

binary

Applies a binary operations to two primitive arrays. This is the fastest way to perform an operation on two primitive array when the benefits of a vectorized operation outweighs the cost of branching nulls and non-nulls.

binary_checked

Version of binary that creates a mutable bitmap that is used to keep track of checked operations. The resulting bitmap is compared with the array bitmap to create the final validity array.

binary_with_bitmap

Version of binary that returns an array and bitmap. Used when working with overflowing operations

try_binary

Version of binary that checks for errors in the closure used to create the buffer

try_unary

Version of unary that checks for errors in the closure used to create the buffer

unary

Applies an unary and infallible function to a PrimitiveArray. This is the fastest way to perform an operation on a PrimitiveArray when the benefits of a vectorized operation outweighs the cost of branching nulls and non-nulls.

unary_checked

Version of unary that creates a mutable bitmap that is used to keep track of checked operations. The resulting bitmap is compared with the array bitmap to create the final validity array.

unary_with_bitmap

Version of unary that returns an array and bitmap. Used when working with overflowing operations