Crate re_viewer::external::re_arrow_store

The Rerun datastore, implemented on top of Apache Arrow using the arrow2 crate.

This crate is an in-memory time series database for Rerun log data. It is indexed by Entity path, component, timeline, and time. It supports out-of-order insertions, and fast O(log(N)) queries.

See DataStore for an overview of the core data structures.

See DataStore::latest_at and DataStore::range for the documentation of the public read APIs.

See DataStore::insert_row for the documentation of the public write APIs.

Feature flags

testing — Enables access to re_types’ test components/datatypes.

deadlock_detection — Enables parking_lot’s deadlock detection background thread.

polars — Integration with polars, to efficiently use the datastore with dataframes.

core_benchmarks_only — When set, only run the core set of benchmark suites. Commonly set implicitly by –all-features, e.g. on CI.

Modules

external

polars_util

Structs

DataStore

A complete data store: covers all timelines, all entities, everything.

DataStoreConfig

DataStoreRowStats

DataStoreStats

Deleted

EntityStats

GarbageCollectionOptions

LatestAtQuery

A query at a given time, for a given timeline.

RangeQuery

A query over a time range, for a given timeline.

StoreGeneration

Incremented on each edit

StreamReader

Arrow Stream reader.

TimeInt

A 64-bit number describing either nanoseconds OR sequence numbers.

TimeRange

Timeline

A time frame/space, e.g. log_time or frame_nr, coupled with the type of time it keeps.

Enums

GarbageCollectionTarget

StreamState

Encodes the stream’s status after each read.

TimeType

The type of a TimeInt or Timeline.

WriteError

Traits

ArrayExt

Type Definitions

WriteResult