Module re_viewer::external::re_arrow_store::external::arrow2::ffi

contains FFI bindings to import and export Array via Arrow’s C Data Interface

Modules

mmap

Functionality to mmap in-memory data regions.

Structs

ArrowArray

ABI-compatible struct for ArrowArray

ArrowArrayStream

ABI-compatible struct for ArrowArrayStream.

ArrowArrayStreamReader

Implements an iterator of Array consumed from the C stream interface.

ArrowSchema

ABI-compatible struct for ArrowSchema

Functions

export_array_to_c

Exports an Box<dyn Array> to the C data interface.

export_field_to_c

Exports a Field to the C data interface.

export_iterator

Exports an iterator to the C stream interface

import_array_from_c⚠

Imports an Array from the C data interface.

import_field_from_c⚠

Imports a Field from the C data interface.