Trait re_viewer::external::re_chunk::external::arrow::array::RecordBatchReader

pub trait RecordBatchReader: Iterator<Item = Result<RecordBatch, ArrowError>> {
    // Required method
    fn schema(&self) -> Arc<Schema>;

    // Provided method
    fn next_batch(&mut self) -> Result<Option<RecordBatch>, ArrowError> { ... }
}

Trait for types that can read RecordBatchโs.

To create from an iterator, see RecordBatchIterator.

Required Methods

fn schema(&self) -> Arc<Schema>

Returns the schema of this RecordBatchReader.

Implementation of this trait should guarantee that all RecordBatchโs returned by this reader should have the same schema as returned from this method.

Provided Methods

fn next_batch(&mut self) -> Result<Option<RecordBatch>, ArrowError>

๐Deprecated since 2.0.0: This method is deprecated in favour of next from the trait Iterator.

Reads the next RecordBatch.