re_viewer::external::re_chunk::external::arrow::buffer
Struct OffsetBuffer

pub struct OffsetBuffer<O>(ScalarBuffer<O>)
where
    O: ArrowNativeType;

A non-empty buffer of monotonically increasing, positive integers.

OffsetBuffer are used to represent ranges of offsets. An OffsetBuffer of N+1 items contains N such ranges. The start offset for element i is offsets[i] and the end offset is offsets[i+1]. Equal offsets represent an empty range.

Example

This example shows how 5 distinct ranges, are represented using a 6 entry OffsetBuffer. The first entry (0, 3) represents the three offsets 0, 1, 2. The entry (3,3) represent no offsets (e.g. an empty list).

  â”Œâ”€â”€â”€â”€â”€â”€â”€â”                â”Œâ”€â”€â”€â”
  â”‚ (0,3) â”‚                â”‚ 0 â”‚
  â”œâ”€â”€â”€â”€â”€â”€â”€â”¤                â”œâ”€â”€â”€â”¤
  â”‚ (3,3) â”‚                â”‚ 3 â”‚
  â”œâ”€â”€â”€â”€â”€â”€â”€â”¤                â”œâ”€â”€â”€â”¤
  â”‚ (3,4) â”‚                â”‚ 3 â”‚
  â”œâ”€â”€â”€â”€â”€â”€â”€â”¤                â”œâ”€â”€â”€â”¤
  â”‚ (4,5) â”‚                â”‚ 4 â”‚
  â”œâ”€â”€â”€â”€â”€â”€â”€â”¤                â”œâ”€â”€â”€â”¤
  â”‚ (5,7) â”‚                â”‚ 5 â”‚
  â””â”€â”€â”€â”€â”€â”€â”€â”˜                â”œâ”€â”€â”€â”¤
                           â”‚ 7 â”‚
                           â””â”€â”€â”€â”˜

                       Offsets Buffer
   Logical
   Offsets

 (offsets[i],
  offsets[i+1])

Tuple Fields
0: ScalarBuffer<O>
Implementations

impl<O> OffsetBuffer<O>
where O: ArrowNativeType,

pub fn new(buffer: ScalarBuffer<O>) -> OffsetBuffer<O>

Create a new OffsetBuffer from the provided ScalarBuffer

Panics

Panics if buffer is not a non-empty buffer containing monotonically increasing values greater than or equal to zero

pub unsafe fn new_unchecked(buffer: ScalarBuffer<O>) -> OffsetBuffer<O>

Create a new OffsetBuffer from the provided ScalarBuffer

Safety

buffer must be a non-empty buffer containing monotonically increasing values greater than or equal to zero

pub fn new_empty() -> OffsetBuffer<O>

Create a new OffsetBuffer containing a single 0 value

pub fn new_zeroed(len: usize) -> OffsetBuffer<O>

Create a new OffsetBuffer containing len + 1 0 values

pub fn from_lengths<I>(lengths: I) -> OffsetBuffer<O>
where I: IntoIterator<Item = usize>,

Create a new OffsetBuffer from the iterator of slice lengths

let offsets = OffsetBuffer::<i32>::from_lengths([1, 3, 5]);
assert_eq!(offsets.as_ref(), &[0, 1, 4, 9]);

Panics

Panics on overflow

pub fn lengths(&self) -> impl ExactSizeIterator

Get an Iterator over the lengths of this OffsetBuffer

let offsets = OffsetBuffer::<_>::new(ScalarBuffer::<i32>::from(vec![0, 1, 4, 9]));
assert_eq!(offsets.lengths().collect::<Vec<usize>>(), vec![1, 3, 5]);

Empty OffsetBuffer will return an empty iterator

let offsets = OffsetBuffer::<i32>::new_empty();
assert_eq!(offsets.lengths().count(), 0);

This can be used to merge multiple OffsetBuffers to one

let buffer1 = OffsetBuffer::<i32>::from_lengths([2, 6, 3, 7, 2]);
let buffer2 = OffsetBuffer::<i32>::from_lengths([1, 3, 5, 7, 9]);

let merged = OffsetBuffer::<i32>::from_lengths(
    vec![buffer1, buffer2].iter().flat_map(|x| x.lengths())
);

assert_eq!(merged.lengths().collect::<Vec<_>>(), &[2, 6, 3, 7, 2, 1, 3, 5, 7, 9]);

pub fn shrink_to_fit(&mut self)

Free up unused memory.

pub fn inner(&self) -> &ScalarBuffer<O>

Returns the inner ScalarBuffer

pub fn into_inner(self) -> ScalarBuffer<O>

Returns the inner ScalarBuffer, consuming self

pub fn slice(&self, offset: usize, len: usize) -> OffsetBuffer<O>

Returns a zero-copy slice of this buffer with length len and starting at offset

pub fn ptr_eq(&self, other: &OffsetBuffer<O>) -> bool

Returns true if this OffsetBuffer is equal to other, using pointer comparisons to determine buffer equality. This is cheaper than PartialEq::eq but may return false when the arrays are logically equal