re_viewer::external::re_chunk::external::arrow::ipc::gen
Module Schema

Structs

Binary

Opaque binary data

BinaryArgs

BinaryBuilder

BinaryView

Logically the same as Binary, but the internal representation uses a view struct that contains the string length and either the string’s entire data inline (for small strings) or an inlined prefix, an index of another buffer, and an offset pointing to a slice in that buffer (for non-small strings).

BinaryViewArgs

BinaryViewBuilder

Bool

BoolArgs

BoolBuilder

Buffer

Date

Date is either a 32-bit or 64-bit signed integer type representing an elapsed time since UNIX epoch (1970-01-01), stored in either of two units:

DateArgs

DateBuilder

DateUnit

Decimal

Exact decimal value represented as an integer value in two’s complement. Currently only 128-bit (16-byte) and 256-bit (32-byte) integers are used. The representation uses the endianness indicated in the Schema.

DecimalArgs

DecimalBuilder

DictionaryEncoding

DictionaryEncodingArgs

DictionaryEncodingBuilder

DictionaryKind

Duration

DurationArgs

DurationBuilder

Endianness

Feature

Represents Arrow Features that might not have full support within implementations. This is intended to be used in two scenarios:

Field

FieldArgs

FieldBuilder

FixedSizeBinary

FixedSizeBinaryArgs

FixedSizeBinaryBuilder

FixedSizeList

FixedSizeListArgs

FixedSizeListBuilder

FloatingPoint

FloatingPointArgs

FloatingPointBuilder

Int

IntArgs

IntBuilder

Interval

IntervalArgs

IntervalBuilder

IntervalUnit

KeyValue

KeyValueArgs

KeyValueBuilder

LargeBinary

Same as Binary, but with 64-bit offsets, allowing to represent extremely large data values.

LargeBinaryArgs

LargeBinaryBuilder

LargeList

Same as List, but with 64-bit offsets, allowing to represent extremely large data values.

LargeListArgs

LargeListBuilder

LargeListView

Same as ListView, but with 64-bit offsets and sizes, allowing to represent extremely large data values.

LargeListViewArgs

LargeListViewBuilder

LargeUtf8

Same as Utf8, but with 64-bit offsets, allowing to represent extremely large data values.

LargeUtf8Args

LargeUtf8Builder

List

ListArgs

ListBuilder

ListView

Represents the same logical types that List can, but contains offsets and sizes allowing for writes in any order and sharing of child values among list values.

ListViewArgs

ListViewBuilder

Map

A Map is a logical nested type that is represented as

MapArgs

MapBuilder

MetadataVersion

Null

These are stored in the flatbuffer in the Type union below

NullArgs

NullBuilder

Precision

RunEndEncoded

Contains two child arrays, run_ends and values. The run_ends child array must be a 16/32/64-bit integer array which encodes the indices at which the run with the value in each corresponding index in the values child array ends. Like list/struct types, the value array can be of any type.

RunEndEncodedArgs

RunEndEncodedBuilder

Schema

SchemaArgs

SchemaBuilder

Struct_

A Struct_ in the flatbuffer metadata is the same as an Arrow Struct (according to the physical memory layout). We used Struct_ here as Struct is a reserved word in Flatbuffers

Struct_Args

Struct_Builder

Time

Time is either a 32-bit or 64-bit signed integer type representing an elapsed time since midnight, stored in either of four units: seconds, milliseconds, microseconds or nanoseconds.

TimeArgs

TimeBuilder

TimeUnit

Timestamp

Timestamp is a 64-bit signed integer representing an elapsed time since a fixed epoch, stored in either of four units: seconds, milliseconds, microseconds or nanoseconds, and is optionally annotated with a timezone.

TimestampArgs

TimestampBuilder

Type

TypeUnionTableOffset

Union

A union is a complex type with children in Field By default ids in the type vector refer to the offsets in the children optionally typeIds provides an indirection between the child offset and the type id for each child typeIds[offset] is the id used in the type vector

UnionArgs

UnionBuilder

UnionMode

Utf8

Unicode with UTF-8 encoding

Utf8Args

Utf8Builder

Utf8View

Logically the same as Utf8, but the internal representation uses a view struct that contains the string length and either the string’s entire data inline (for small strings) or an inlined prefix, an index of another buffer, and an offset pointing to a slice in that buffer (for non-small strings).

Utf8ViewArgs

Utf8ViewBuilder

Enums

BinaryOffset

BinaryViewOffset

BoolOffset

DateOffset

DecimalOffset

DictionaryEncodingOffset

DurationOffset

FieldOffset

FixedSizeBinaryOffset

FixedSizeListOffset

FloatingPointOffset

IntOffset

IntervalOffset

KeyValueOffset

LargeBinaryOffset

LargeListOffset

LargeListViewOffset

LargeUtf8Offset

ListOffset

ListViewOffset

MapOffset

NullOffset

RunEndEncodedOffset

SchemaOffset

Struct_Offset

TimeOffset

TimestampOffset

UnionOffset

Utf8Offset

Utf8ViewOffset

Constants

ENUM_MAX_DATE_UNITDeprecated

ENUM_MAX_DICTIONARY_KINDDeprecated

ENUM_MAX_ENDIANNESSDeprecated

ENUM_MAX_FEATUREDeprecated

ENUM_MAX_INTERVAL_UNITDeprecated

ENUM_MAX_METADATA_VERSIONDeprecated

ENUM_MAX_PRECISIONDeprecated

ENUM_MAX_TIME_UNITDeprecated

ENUM_MAX_TYPEDeprecated

ENUM_MAX_UNION_MODEDeprecated

ENUM_MIN_DATE_UNITDeprecated

ENUM_MIN_DICTIONARY_KINDDeprecated

ENUM_MIN_ENDIANNESSDeprecated

ENUM_MIN_FEATUREDeprecated

ENUM_MIN_INTERVAL_UNITDeprecated

ENUM_MIN_METADATA_VERSIONDeprecated

ENUM_MIN_PRECISIONDeprecated

ENUM_MIN_TIME_UNITDeprecated

ENUM_MIN_TYPEDeprecated

ENUM_MIN_UNION_MODEDeprecated

ENUM_VALUES_DATE_UNITDeprecated

ENUM_VALUES_DICTIONARY_KINDDeprecated

ENUM_VALUES_ENDIANNESSDeprecated

ENUM_VALUES_FEATUREDeprecated

ENUM_VALUES_INTERVAL_UNITDeprecated

ENUM_VALUES_METADATA_VERSIONDeprecated

ENUM_VALUES_PRECISIONDeprecated

ENUM_VALUES_TIME_UNITDeprecated

ENUM_VALUES_TYPEDeprecated

ENUM_VALUES_UNION_MODEDeprecated

Functions

finish_schema_buffer

finish_size_prefixed_schema_buffer

root_as_schema

Verifies that a buffer of bytes contains a Schema and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_schema_unchecked.

root_as_schema_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a Schema and returns it.

root_as_schema_with_opts

Verifies, with the given options, that a buffer of bytes contains a Schema and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_schema_unchecked.

size_prefixed_root_as_schema

Verifies that a buffer of bytes contains a size prefixed Schema and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_schema_unchecked.

size_prefixed_root_as_schema_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a size prefixed Schema and returns it.

size_prefixed_root_as_schema_with_opts

Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed Schema and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_schema_unchecked.