re_viewer::external::re_chunk::external::re_log_types::external::re_types_core
Trait View

pub trait View {
    // Required method
    fn identifier() -> ViewClassIdentifier;
}

Views are the panels shown in the viewer’s viewport and the primary means of inspecting & visualizing previously logged data.

In addition to the data that it contains via ViewContents, each view has several view properties that configure how it behaves. Each view property is a crate::Archetype that is stored in the viewer’s blueprint database.

Required Methods

fn identifier() -> ViewClassIdentifier

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.