Module re_viewer::external::re_chunk_store::external::arrow::compute::kernels::take

Defines take kernel for Array

Structs

TakeOptions

Options that define how take should behave

Functions

take

Take elements by index from Array, creating a new Array from those indexes.

take_arrays

For each ArrayRef in the Vec<ArrayRef>, take elements by index and create a new Vec<ArrayRef> from those indices.

take_record_batch

Take rows by index from RecordBatch and returns a new RecordBatch from those indexes.