re_viewer::external::re_chunk_store::external::arrow::util::bit_util
Function unset_bit_raw

pub unsafe fn unset_bit_raw(data: *mut u8, i: usize)

Sets bit at position i for data to 0

Safety

Note this doesn’t do any bound checking, for performance reason. The caller is responsible to guarantee that i is within bounds.