Module re_viewer::external::re_chunk_store::external::arrow2::ffi::mmap

Functionality to mmap in-memory data regions.

Functions

bitmap⚠

Creates a (non-null) BooleanArray from a slice of bits. This does not have memcopy and is the fastest way to create a BooleanArray.

slice⚠

Creates a (non-null) PrimitiveArray from a slice of values. This does not have memcopy and is the fastest way to create a PrimitiveArray.