Trait re_viewer::external::re_chunk_store::external::re_chunk::external::re_log_types::external::re_types_core::View

pub trait View {
    // Required method
    fn identifier() -> SpaceViewClassIdentifier;
}

Views are the panels shown in the viewer’s viewport and the primary means of inspecting & visualizing previously logged data.

In addition to the data that it contains via SpaceViewContents, each view has several view properties that configure how it behaves. Each view property is a crate::Archetype that is stored in the viewer’s blueprint database.

Required Methods

fn identifier() -> SpaceViewClassIdentifier

Object Safety

This trait is not object safe.