Struct re_viewer::external::re_chunk_store::external::re_chunk::UnitChunkShared

pub struct UnitChunkShared(Arc<Chunk>);

A ChunkShared that is guaranteed to always contain a single row’s worth of data.

Tuple Fields
0: Arc<Chunk>
Implementations

impl UnitChunkShared

pub fn into_chunk(self) -> Arc<Chunk>

impl UnitChunkShared

pub fn index(&self, timeline: &Timeline) -> Option<(TimeInt, RowId)>

Returns the index ((TimeInt, RowId) pair) of the single row within, on the given timeline.

Returns the single static index if the chunk is static.

pub fn row_id(&self) -> Option<RowId>

Returns the RowId of the single row within, on the given timeline.

Returns the single static RowId if the chunk is static.

pub fn num_instances(&self, component_name: &ComponentName) -> u64

Returns the number of instances of the single row within for a given component.

impl UnitChunkShared

pub fn component_batch_raw( &self, component_name: &ComponentName, ) -> Option<Box<dyn Array>>

Returns the raw data for the specified component.

pub fn component_batch<C>(&self) -> Option<Result<Vec<C>, ChunkError>>
where C: Component,

Returns the deserialized data for the specified component.

Returns an error if the data cannot be deserialized.

pub fn component_instance_raw( &self, component_name: &ComponentName, instance_index: usize, ) -> Option<Result<Box<dyn Array>, ChunkError>>

Returns the raw data for the specified component at the given instance index.

Returns an error if the instance index is out of bounds.

pub fn component_instance<C>( &self, instance_index: usize, ) -> Option<Result<C, ChunkError>>
where C: Component,

Returns the deserialized data for the specified component at the given instance index.

Returns an error if the data cannot be deserialized, or if the instance index is out of bounds.

pub fn component_mono_raw( &self, component_name: &ComponentName, ) -> Option<Result<Box<dyn Array>, ChunkError>>

Returns the raw data for the specified component, assuming a mono-batch.

Returns an error if the underlying batch is not of unit length.

pub fn component_mono<C>(&self) -> Option<Result<C, ChunkError>>
where C: Component,

Returns the deserialized data for the specified component, assuming a mono-batch.

Returns an error if the data cannot be deserialized, or if the underlying batch is not of unit length.