Crate re_viewer::external::re_data_store

This is how we store and index logging data.

Feature flags

serde — Enable (de)serialization using serde.

Modules

entity_properties

entity_tree

store_db

Structs

ActuallyDeleted

Book-keeping required after a GC purge to keep track of what was removed from children, so it can also be removed from the parents.

ComponentName

The fully-qualified name of a Component, e.g. rerun.components.Position2D.

ComponentStats

EntityPath

The unique identifier of an entity, e.g. camera/"ACME Örnöga"/points

EntityProperties

EntityPropertyMap

Properties for a collection of entities.

EntityTree

Tree of entity paths, plus components at the leaves.

ExtraQueryHistory

When showing an entity in the history view, add this much history to it.

InstancePath

The path to either a specific instance of an entity, or the whole entity (splat).

InstancePathHash

Hashes of the components of an InstancePath.

StoreDb

A in-memory database built from a stream of LogMsges.

TimeHistogramPerTimeline

Number of messages per time per timeline.

TimeInt

A 64-bit number describing either nanoseconds OR sequence numbers.

Timeline

A time frame/space, e.g. log_time or frame_nr, coupled with the type of time it keeps.

TimesPerTimeline

Number of messages per time per timeline.

VersionedInstancePath

A versioned path (i.e. pinned to a specific RowId) to either a specific instance of an entity, or the whole entity (splat).

VersionedInstancePathHash

Hashes of the components of a VersionedInstancePath.

Enums

ColorMapper

Colormap

EditableAutoValue

A value that is either determined automatically by some heuristic, or specified by the user.

EntityPathPart

The different parts that make up an EntityPath.

Error

The errors that can occur when misusing the data store.

Index

The key of a table, or an array index.

TimeQuery

A query in time.

Type Definitions

Result

TimeHistogram

Number of messages per time