Trait re_viewer::external::re_data_store::external::arrow2::array::DictionaryKey

pub unsafe trait DictionaryKey: NativeType + TryInto<usize> + TryFrom<usize> + Hash {
    const KEY_TYPE: IntegerType;

    // Provided methods
    unsafe fn as_usize(self) -> usize { ... }
    fn always_fits_usize() -> bool { ... }
}

Trait denoting NativeTypes that can be used as keys of a dictionary.

Safety

Any implementation of this trait must ensure that always_fits_usize only returns true if all values succeeds on value::try_into::<usize>().unwrap().

Required Associated Constants

const KEY_TYPE: IntegerType

The corresponding IntegerType of this key

Provided Methods

unsafe fn as_usize(self) -> usize

Represents this key as a usize.

Safety

The caller must have checked that the value can be casted to usize.

fn always_fits_usize() -> bool

If the key type always can be converted to usize.

Object Safety

This trait is not object safe.