Function re_viewer::external::re_entity_db::external::re_query::clamped_zip::clamped_zip_1x4

pub fn clamped_zip_1x4<R0, O0, O1, O2, O3, D0, D1, D2, D3>(
    r0: R0,
    o0: O0,
    o0_default_fn: D0,
    o1: O1,
    o1_default_fn: D1,
    o2: O2,
    o2_default_fn: D2,
    o3: O3,
    o3_default_fn: D3
) -> ClampedZip1x4<<R0 as IntoIterator>::IntoIter, <O0 as IntoIterator>::IntoIter, <O1 as IntoIterator>::IntoIter, <O2 as IntoIterator>::IntoIter, <O3 as IntoIterator>::IntoIter, D0, D1, D2, D3> 
where
    R0: IntoIterator,
    O0: IntoIterator,
    <O0 as IntoIterator>::Item: Clone,
    O1: IntoIterator,
    <O1 as IntoIterator>::Item: Clone,
    O2: IntoIterator,
    <O2 as IntoIterator>::Item: Clone,
    O3: IntoIterator,
    <O3 as IntoIterator>::Item: Clone,
    D0: Fn() -> <O0 as IntoIterator>::Item,
    D1: Fn() -> <O1 as IntoIterator>::Item,
    D2: Fn() -> <O2 as IntoIterator>::Item,
    D3: Fn() -> <O3 as IntoIterator>::Item,

Returns a new ClampedZip1x4 iterator.

The number of elements in a clamped zip iterator corresponds to the number of elements in the shortest of its required iterators (r0).

Optional iterators (o0, o1, o2, o3) will repeat their latest values if they happen to be too short to be zipped with the shortest of the required iterators.

If an optional iterator is not only too short but actually empty, its associated default function (o0_default_fn, o1_default_fn, o2_default_fn, o3_default_fn) will be executed and the resulting value repeated as necessary.