Struct re_viewer::external::re_entity_db::external::re_query::LatestAtResults

pub struct LatestAtResults {
    pub compound_index: (TimeInt, RowId),
    pub components: HashMap<ComponentName, Arc<LatestAtComponentResults>, BuildHasherDefault<NoHashHasher<ComponentName>>>,
}

Results for a latest-at query.

The data is both deserialized and resolved/converted.

Use LatestAtResults::get, LatestAtResults::get_required and LatestAtResults::get_or_empty in order to access the results for each individual component.

Fields
compound_index: (TimeInt, RowId)

The compound index of this query result.

A latest-at query is a compound operation that gathers data from many different rows. The index of that compound result corresponds to the index of most the recent row in all the sub-results, as defined by time and row-id order.

components: HashMap<ComponentName, Arc<LatestAtComponentResults>, BuildHasherDefault<NoHashHasher<ComponentName>>>

Results for each individual component.

Implementations

impl LatestAtResults

pub fn contains(&self, component_name: impl Into<ComponentName>) -> bool

pub fn get( &self, component_name: impl Into<ComponentName> ) -> Option<&LatestAtComponentResults>

Returns the LatestAtComponentResults for the specified Component.

pub fn get_required( &self, component_name: impl Into<ComponentName> ) -> Result<&LatestAtComponentResults, QueryError>

Returns the LatestAtComponentResults for the specified Component.

Returns an error if the component is not present.

pub fn get_or_empty( &self, component_name: impl Into<ComponentName> ) -> &LatestAtComponentResults

Returns the LatestAtComponentResults for the specified Component.

Returns empty results if the component is not present.