Struct re_viewer::external::re_log_types::NonMinI64

#[repr(transparent)]
pub struct NonMinI64(NonZeroI64);

An integer that is known not to equal its minimum value.

Tuple Fields
0: NonZeroI64
Implementations

impl NonMinI64

pub const ZERO: NonMinI64 = _

pub const ONE: NonMinI64 = _

pub const MIN: NonMinI64 = _

pub const MAX: NonMinI64 = _

pub const fn new(value: i64) -> Option<NonMinI64>

Creates a new non-min if the given value is not the minimum value.

pub const unsafe fn new_unchecked(value: i64) -> NonMinI64

Creates a new non-min without checking the value.

Safety

The value must not equal the minimum representable value for the primitive type.

pub const fn get(&self) -> i64

Returns the value as a primitive type.