Crate re_viewer::external::re_space_view

Rerun Space View utilities

Types & utilities for defining Space View classes and communicating with the Viewport.

Modules

controls

external

Structs

HybridLatestAtResults

Wrapper that contains the results of a latest-at query with possible overrides.

HybridResultsChunkIter

The iterator type backing HybridResults::iter_as.

Enums

HybridResults

ScreenshotMode

Traits

DataResultQuery

RangeResultsExt

Extension traits to abstract query result handling for all spatial space views.

Functions

diff_component_filter

Utility for implementing re_viewer_context::VisualizerAdditionalApplicabilityFilter using on the properties of a concrete component.

latest_at_with_blueprint_resolved_data

Queries for the given component_names using latest-at semantics with blueprint support.

range_with_blueprint_resolved_data

Queries for the given component_names using range semantics with blueprint support.

suggest_space_view_for_each_entity

Spawns a space view for each single entity which is visualizable & indicator-matching for a given visualizer.

view_property_ui

Display the UI for editing all components of a blueprint archetype.