Crate re_viewer::external::re_space_view::external::re_chunk_store::external::re_chunk::external::re_log_types

The different types that make up the rerun log format.

Feature flags

serde — Enable (de)serialization using serde.

Mono-components

Some components, mostly transform related ones, are “mono-components”. This means that Rerun makes assumptions that depend on this component only taking on a singular value for all instances of an Entity. Where possible, exposed APIs will force these components to be logged as a singular instance. However, it is an error with undefined behavior to manually use lower-level APIs to log a batched mono-component.

This requirement is especially apparent with transforms: Each entity must have a unique transform chain, e.g. the entity foo/bar/baz is has the transform that is the product of foo.transform * foo/bar.transform * foo/bar/baz.transform.

Modules

arrow_msg

ArrowMsg is the crate::LogMsg sub-type containing an Arrow payload.

example_components

Example components to be used for tests and docs

external

hash

path

Every logged entity in Rerun is logged to an EntityPath.

time_point

Macros

entity_path

Build an EntityPath from parts that are not escaped:

entity_path_vec

Build a Vec<EntityPathPart>:

impl_into_enum

Structs

ApplicationId

The user-chosen name of the application doing the logging.

ArrowChunkReleaseCallback

An arbitrary callback to be run when an ArrowMsg, and more specifically the Arrow2Chunk within it, goes out of scope.

ArrowMsg

Message containing an Arrow payload

BlueprintActivationCommand

Command used for activating a blueprint once it has been fully transmitted.

ComponentPath

A EntityPath plus a ComponentName.

DataPath

A general path to some data.

Duration

A signed duration represented as nanoseconds since unix epoch

EntityPath

The unique identifier of an entity, e.g. camera/3/points

EntityPathFilter

A way to filter a set of EntityPaths.

EntityPathHash

A 64 bit hash of EntityPath with very small risk of collision.

EntityPathPart

The different parts that make up an EntityPath.

EntityPathRule

EntityPathSubs

A set of substitutions for entity paths.

Instance

A unique numeric index for each individual instance within a batch.

NonMinI64

An integer that is known not to equal its minimum value.

PythonVersion

ResolvedTimeRange

ResolvedTimeRangeF

Like ResolvedTimeRange, but using TimeReal for improved precision.

SetStoreInfo

StoreId

A unique id per store.

StoreInfo

Information about a recording or blueprint.

Time

A date-time represented as nanoseconds since unix epoch

TimeInt

A 64-bit number describing either nanoseconds, sequence numbers or fully static data.

TimePoint

A point in time on any number of Timelines.

TimeReal

Either nanoseconds or sequence numbers.

Timeline

A time frame/space, e.g. log_time or frame_nr, coupled with the type of time it keeps.

TimelineName

The name of a timeline. Often something like "log_time" or "frame_nr".

TryFromIntError

An error type returned when a checked integral type conversion fails (mimics std::num::TryFromIntError)

Enums

FileSource

LogMsg

The most general log message sent from the SDK to the server.

PathParseError

RuleEffect

StoreKind

What kind of Store this is.

StoreSource

The source of a recording or blueprint.

TimeType

The type of a TimeInt or Timeline.

TimeZone

Traits

VecDequeInsertionExt

Extends VecDeque with extra insertion routines.

VecDequeRemovalExt

Extends VecDeque with extra removal routines.

VecDequeSortingExt

Extends VecDeque with extra sorting routines.

Functions

build_frame_nr

Build a (Timeline, TimeInt) tuple from frame_nr suitable for inserting in a TimePoint.

build_log_time

Build a (Timeline, TimeInt) tuple from log_time suitable for inserting in a TimePoint.