Module re_viewer::external::re_space_view::external::re_chunk_store::external::re_chunk::external::re_log_types::arrow_msg

ArrowMsg is the crate::LogMsg sub-type containing an Arrow payload.

We have custom implementations of serde::Serialize and serde::Deserialize that wraps the inner Arrow serialization of Arrow2Schema and Arrow2Chunk.

Structs

ArrowChunkReleaseCallback

An arbitrary callback to be run when an ArrowMsg, and more specifically the Arrow2Chunk within it, goes out of scope.

ArrowMsg

Message containing an Arrow payload