Function re_viewer::external::re_space_view::external::re_chunk_store::external::re_chunk::util::filter_array

pub fn filter_array<A>(array: &A, filter: &BooleanArray) -> A
where
    A: Array + Clone,

Applies a filter kernel to the given array.

Panics iff the length of the filter doesn’t match the length of the array.

In release builds, filters are allowed to have null entries (they will be interpreted as false). In debug builds, null entries will panic.

Note: a filter kernel copies the data in order to make the resulting arrays contiguous in memory.

Takes care of up- and down-casting the data back and forth on behalf of the caller.