Struct re_viewer::external::re_space_view::external::re_data_store::external::arrow2::array::BooleanArray

pub struct BooleanArray {
    data_type: DataType,
    values: Bitmap,
    validity: Option<Bitmap>,
}

A BooleanArray is Arrow’s semantically equivalent of an immutable Vec<Option<bool>>. It implements Array.

One way to think about a BooleanArray is (DataType, Arc<Vec<u8>>, Option<Arc<Vec<u8>>>) where:

the first item is the array’s logical type

the second is the immutable values

the third is the immutable validity (whether a value is null or not as a bitmap).

The size of this struct is O(1), as all data is stored behind an std::sync::Arc.

Example

use re_arrow2::array::BooleanArray;
use re_arrow2::bitmap::Bitmap;
use re_arrow2::buffer::Buffer;

let array = BooleanArray::from([Some(true), None, Some(false)]);
assert_eq!(array.value(0), true);
assert_eq!(array.iter().collect::<Vec<_>>(), vec![Some(true), None, Some(false)]);
assert_eq!(array.values_iter().collect::<Vec<_>>(), vec![true, false, false]);
// the underlying representation
assert_eq!(array.values(), &Bitmap::from([true, false, false]));
assert_eq!(array.validity(), Some(&Bitmap::from([true, false, true])));

Fields
data_type: DataTypevalues: Bitmapvalidity: Option<Bitmap>
Implementations

impl BooleanArray

pub fn try_new( data_type: DataType, values: Bitmap, validity: Option<Bitmap> ) -> Result<BooleanArray, Error>

The canonical method to create a BooleanArray out of low-end APIs.

Errors

This function errors iff:

The validity is not None and its length is different from values’s length

The data_type’s PhysicalType is not equal to PhysicalType::Boolean.

pub fn new( data_type: DataType, values: Bitmap, validity: Option<Bitmap> ) -> BooleanArray

Alias to Self::try_new().unwrap()

pub fn iter(&self) -> ZipValidity<bool, BitmapIter<'_>, BitmapIter<'_>>

Returns an iterator over the optional values of this BooleanArray.

pub fn values_iter(&self) -> BitmapIter<'_>

Returns an iterator over the values of this BooleanArray.

pub fn len(&self) -> usize

Returns the length of this array

pub fn values(&self) -> &Bitmap

The values Bitmap. Values on null slots are undetermined (they can be anything).

pub fn validity(&self) -> Option<&Bitmap>

Returns the optional validity.

pub fn data_type(&self) -> &DataType

Returns the arrays’ DataType.

pub fn value(&self, i: usize) -> bool

Returns the value at index i

Panic

This function panics iff i >= self.len().

pub unsafe fn value_unchecked(&self, i: usize) -> bool

Returns the element at index i as bool

Safety

Caller must be sure that i < self.len()

pub fn get(&self, i: usize) -> Option<bool>

Returns the element at index i or None if it is null

Panics

iff i >= self.len()

pub fn slice(&mut self, offset: usize, length: usize)

Slices this BooleanArray.

Implementation

This operation is O(1) as it amounts to increase up to two ref counts.

Panic

This function panics iff offset + length > self.len().

pub unsafe fn slice_unchecked(&mut self, offset: usize, length: usize)

Slices this BooleanArray.

Implementation

This operation is O(1) as it amounts to increase two ref counts.

Safety

The caller must ensure that offset + length <= self.len().

pub fn sliced(self, offset: usize, length: usize) -> BooleanArray

Returns this array sliced.

Implementation

This function is O(1).

Panics

iff offset + length > self.len().

pub unsafe fn sliced_unchecked( self, offset: usize, length: usize ) -> BooleanArray

Returns this array sliced.

Implementation

This function is O(1).

Safety

The caller must ensure that offset + length <= self.len().

pub fn with_validity(self, validity: Option<Bitmap>) -> BooleanArray

Returns this array with a new validity.

Panic

Panics iff validity.len() != self.len().

pub fn set_validity(&mut self, validity: Option<Bitmap>)

Sets the validity of this array.

Panics

This function panics iff values.len() != self.len().

pub fn boxed(self) -> Box<dyn Array>

Boxes this array into a Box<dyn Array>.

pub fn arced(self) -> Arc<dyn Array>

Arcs this array into a std::sync::Arc<dyn Array>.

pub fn with_values(&self, values: Bitmap) -> BooleanArray

Returns a clone of this BooleanArray with new values.

Panics

This function panics iff values.len() != self.len().

pub fn set_values(&mut self, values: Bitmap)

Sets the values of this BooleanArray.

Panics

This function panics iff values.len() != self.len().

pub fn apply_values_mut<F>(&mut self, f: F)
where F: Fn(&mut MutableBitmap),

Applies a function f to the values of this array, cloning the values iff they are being shared with others

This is an API to use clone-on-write

Implementation

This function is O(f) if the data is not being shared, and O(N) + O(f) if it is being shared (since it results in a O(N) memcopy).

Panics

This function panics if the function modifies the length of the MutableBitmap.

pub fn into_mut(self) -> Either<BooleanArray, MutableBooleanArray>

Try to convert this BooleanArray to a MutableBooleanArray

pub fn new_empty(data_type: DataType) -> BooleanArray

Returns a new empty BooleanArray.

pub fn new_null(data_type: DataType, length: usize) -> BooleanArray

Returns a new BooleanArray whose all slots are null / None.

pub fn from_trusted_len_values_iter<I>(iterator: I) -> BooleanArray
where I: TrustedLen<Item = bool>,

Creates a new BooleanArray from an TrustedLen of bool.

pub unsafe fn from_trusted_len_values_iter_unchecked<I>( iterator: I ) -> BooleanArray
where I: Iterator<Item = bool>,

Creates a new BooleanArray from an TrustedLen of bool. Use this over BooleanArray::from_trusted_len_iter when the iterator is trusted len but this crate does not mark it as such.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_slice<P>(slice: P) -> BooleanArray
where P: AsRef<[bool]>,

Creates a new BooleanArray from a slice of bool.

pub unsafe fn from_trusted_len_iter_unchecked<I, P>(iterator: I) -> BooleanArray
where P: Borrow<bool>, I: Iterator<Item = Option<P>>,

Creates a BooleanArray from an iterator of trusted length. Use this over BooleanArray::from_trusted_len_iter when the iterator is trusted len but this crate does not mark it as such.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> BooleanArray
where P: Borrow<bool>, I: TrustedLen<Item = Option<P>>,

Creates a BooleanArray from a TrustedLen.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>( iterator: I ) -> Result<BooleanArray, E>
where P: Borrow<bool>, I: Iterator<Item = Result<Option<P>, E>>,

Creates a BooleanArray from an falible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>( iterator: I ) -> Result<BooleanArray, E>
where P: Borrow<bool>, I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a BooleanArray from a TrustedLen.

pub fn into_inner(self) -> (DataType, Bitmap, Option<Bitmap>)

Returns its internal representation

pub unsafe fn from_inner_unchecked( data_type: DataType, values: Bitmap, validity: Option<Bitmap> ) -> BooleanArray

Creates a [BooleanArray] from its internal representation. This is the inverted from [BooleanArray::into_inner]

Safety

Callers must ensure all invariants of this struct are upheld.