Crate re_viewer::external::re_space_view::external::re_query_cache2

Caching datastructures for re_query.

Modules

external

Structs

CacheKey

Uniquely identifies cached query results in the Caches.

CachedLatestAtComponentResults

Lazily cached results for a particular component when using a cached latest-at query.

CachedLatestAtMonoResult

CachedLatestAtResults

Cached results for a latest-at query.

Caches

ClampedZip1x1

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x2

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x3

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x4

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x5

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x6

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x7

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x8

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x9

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x1

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x2

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x3

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x4

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x5

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x6

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x7

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x8

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x9

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

FlatVecDeque

A double-ended queue implemented with a pair of growable ring buffers, where every single entry is a flattened array of values.

Promise

A Promise turns a source DataCell into a new DataCell with the helper of a PromiseResolver.

PromiseId

Uniquely identifies a Promise.

PromiseResolver

Resolves and keeps track of Promises.

Enums

PromiseResult

The result of resolving a Promise through a PromiseResolver.

QueryError

Traits

ErasedFlatVecDeque

A FlatVecDeque that can be erased into a trait object.

ToArchetype

Helper extension trait to convert query results into re_types_core::Archetypes.

Functions

clamped_zip_1x1

Returns a new ClampedZip1x1 iterator.

clamped_zip_1x2

Returns a new ClampedZip1x2 iterator.

clamped_zip_1x3

Returns a new ClampedZip1x3 iterator.

clamped_zip_1x4

Returns a new ClampedZip1x4 iterator.

clamped_zip_1x5

Returns a new ClampedZip1x5 iterator.

clamped_zip_1x6

Returns a new ClampedZip1x6 iterator.

clamped_zip_1x7

Returns a new ClampedZip1x7 iterator.

clamped_zip_1x8

Returns a new ClampedZip1x8 iterator.

clamped_zip_1x9

Returns a new ClampedZip1x9 iterator.

clamped_zip_2x1

Returns a new ClampedZip2x1 iterator.

clamped_zip_2x2

Returns a new ClampedZip2x2 iterator.

clamped_zip_2x3

Returns a new ClampedZip2x3 iterator.

clamped_zip_2x4

Returns a new ClampedZip2x4 iterator.

clamped_zip_2x5

Returns a new ClampedZip2x5 iterator.

clamped_zip_2x6

Returns a new ClampedZip2x6 iterator.

clamped_zip_2x7

Returns a new ClampedZip2x7 iterator.

clamped_zip_2x8

Returns a new ClampedZip2x8 iterator.

clamped_zip_2x9

Returns a new ClampedZip2x9 iterator.

Type Aliases

Result