Module re_viewer::external::re_types::blueprint::archetypes

The blueprint-specific archetypes.

Structs

Background

Archetype: Configuration for the background of a view.

ContainerBlueprint

Archetype: The description of a container.

DataframeQuery

Archetype: The query for the dataframe view.

ForceCenter

Archetype: Tries to move the center of mass of the graph to the origin.

ForceCollisionRadius

Archetype: Resolves collisions between the bounding circles, according to the radius of the nodes.

ForceLink

Archetype: Aims to achieve a target distance between two nodes that are connected by an edge.

ForceManyBody

Archetype: A force between each pair of nodes that ressembles an electrical charge.

ForcePosition

Archetype: Similar to gravity, this force pulls nodes towards a specific position.

LineGrid3D

Archetype: Configuration for the 3D line grid.

MapBackground

Archetype: Configuration for the background map of the map view.

MapZoom

Archetype: Configuration of the map view zoom level.

NearClipPlane

Archetype: Controls the distance to the near clip plane in 3D scene units.

PanelBlueprint

Archetype: Shared state for the 3 collapsible panels.

PlotLegend

Archetype: Configuration for the legend of a plot.

ScalarAxis

Archetype: Configuration for the scalar axis of a plot.

TensorScalarMapping

Archetype: Configures how tensor scalars are mapped to color.

TensorSliceSelection

Archetype: Specifies a 2D slice of a tensor.

TensorViewFit

Archetype: Configures how a selected tensor slice is shown on screen.

ViewBlueprint

Archetype: The description of a single view.

ViewContents

Archetype: The contents of a View.

ViewportBlueprint

Archetype: The top-level description of the viewport.

VisibleTimeRanges

Archetype: Configures what range of each timeline is shown on a view.

VisualBounds2D

Archetype: Controls the visual bounds of a 2D view.