re_viewer::external::re_types::blueprint
Module views

The blueprint-specific views.

Structs

BarChartView

View: A bar chart view.

DataframeView

View: A view to display any data in a tabular form.

GraphView

View: A graph view to display time-variying, directed or undirected graph visualization.

MapView

View: A 2D map view to display geospatial primitives.

Spatial2DView

View: For viewing spatial 2D data.

Spatial3DView

View: For viewing spatial 3D data.

TensorView

View: A view on a tensor of any dimensionality.

TextDocumentView

View: A view of a single text document, for use with archetypes::TextDocument.

TextLogView

View: A view of a text log, for use with archetypes::TextLog.

TimeSeriesView

View: A time series view for scalars over time, for use with archetypes::Scalars.