re_viewer::external::re_types::external::image::imageops
Function crop_imm

pub fn crop_imm<I>(
    image: &I,
    x: u32,
    y: u32,
    width: u32,
    height: u32,
) -> SubImage<&I>
where
    I: GenericImageView,

Return an immutable view into an image The coordinates set the position of the top left corner of the crop.