re_viewer::external::re_types::external::ndarray::prelude
Type Alias RawArrayViewMut

pub type RawArrayViewMut<A, D> = ArrayBase<RawViewRepr<*mut A>, D>;

A mutable array view without a lifetime.

This is similar to ArrayViewMut but does not carry any lifetime or ownership information, and its data cannot be read/written without an unsafe conversion into an ArrayViewMut. The relationship between RawArrayViewMut and ArrayViewMut is somewhat analogous to the relationship between *mut T and &mut T, but RawArrayViewMut has additional requirements that *mut T does not, such as non-nullness.

The RawArrayViewMut<A, D> is parameterized by A for the element type and D for the dimensionality.

Raw array views have all the methods of an array (see ArrayBase).

See also RawArrayView.

Warning

You can’t use this type with an arbitrary raw pointer; see from_shape_ptr for details.

Aliased Type

struct RawArrayViewMut<A, D> { /* private fields */ }