re_viewer::external::re_types::external::ndarray
Trait RawData

pub unsafe trait RawData: Sized {
    type Elem;
}

Array representation trait.

For an array that meets the invariants of the ArrayBase type. This trait does not imply any ownership or lifetime; pointers to elements in the array may not be safe to dereference.

Note: RawData is not an extension interface at this point. Traits in Rust can serve many different roles. This trait is public because it is used as a bound on public methods.

Required Associated Types

type Elem

The array element type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.