Module re_viewer::external::re_types::external::re_types_core::external::arrow::util::bit_iterator

Types for iterating over packed bitmasks

Structs

BitIndexIterator

An iterator of usize whose index in a provided bitmask is true

BitIterator

Iterator over the bits within a packed bitmask

BitSliceIterator

Iterator of contiguous ranges of set bits within a provided packed bitmask

Functions

try_for_each_valid_idx

Calls the provided closure for each index in the provided null mask that is set, using an adaptive strategy based on the null count