re_viewer::external
Crate re_ui

Rerun GUI theme and helpers, built around egui.

Modules

drag_and_drop

Helpers for drag and drop support for reordering hierarchical lists.

filter_widget

icons

list_item

Second-generation list item.

modal

notifications

syntax_highlighting

Macros

icon_text

Create an IconText with the given items.

Structs

ColorTable

A table mapping all combination of Hue and Scale to a egui::Color32.

ColorToken

A token representing a color in the global color table.

CommandPalette

ControlRow

A single row in the controls table.

DesignTokens

The look and feel of the UI.

Help

A help popup where you can show markdown text and controls as a table.

Icon

IconText

Helper to show text with icons in a row. Usually created via the crate::icon_text! macro.

KeyMarkdown

Newtype over [egui::Key] which provides a Display implementation suitable for markdown.

ModifiersMarkdown

Newtype over [egui::Modifiers] which provides a Display implementation suitable for markdown.

MouseButtonMarkdown

Newtype over [egui::PointerButton] which provides a Display implementation suitable for markdown.

MouseButtonText

Helper to show mouse buttons as text/icons.

SectionCollapsingHeader

A collapsible section header, with support for optional help tooltip and button.

TimeDragValue

Drag value widget for editing time values for both sequence and temporal timelines.

TopBarStyle

Enums

Hue

A hue for a ColorToken.

IconTextItem

LabelStyle

The style of a label.

Scale

A color scale for a ColorToken.

UICommand

All the commands we support.

UiLayout

Specifies the context in which the UI is used and the constraints it should follow.

Constants

CUSTOM_WINDOW_DECORATIONS

If true, we hide the native window decoration (the top bar with app title, close button etc), and instead paint our own close/maximize/minimize buttons.

FULLSIZE_CONTENT

If true, we fill the entire window, except for the close/maximize/minimize buttons in the top-left. See https://github.com/emilk/egui/pull/2049

INFO_COLOR

NATIVE_WINDOW_BAR

If true, we show the native window decorations/chrome with the close/maximize/minimize buttons and app title.

SUCCESS_COLOR

Traits

ContextExt

Extension trait for [egui::Context].

SyntaxHighlighting

UICommandSender

Interface for sending UICommand messages.

UiExt

Rerun custom extensions to [egui::Ui].

Functions

apply_style_and_install_loaders

Apply the Rerun design tokens to the given egui context and install image loaders.

arrow_ui

design_tokens_of

Return a reference to the global design tokens structure.

maybe_plus

Shows a “+” if the OS is not Mac. Useful if you want to e.g. show an icon after a modifiers_text

modifiers_text

Helper to add [egui::Modifiers] as text with icons. Will automatically show Cmd/Ctrl based on the OS.

shortcut_with_icon

Shows a keyboard shortcut with a modifier and the given icon.