re_viewer::external::re_ui
Module list_item

Second-generation list item.

TODO(ab): provide some top-level documentation here.

Structs

ButtonContent

Simple ListItemContent to easily display a button in a crate::list_item::ListItem-based UI.

ContentContext

Context provided to ListItemContent implementations

CustomContent

ListItemContent that mostly delegates to a closure.

DebugContent

ListItemContent that displays the content rect.

ItemActionButton

An super::ItemButton that acts as an action button.

ItemMenuButton

An super::ItemButton that acts as a popup menu.

LabelContent

ListItemContent that displays a simple label with optional icon and buttons.

LayoutInfo

Layout information prepared by list_item_scope to be used by super::ListItemContent.

ListItem

Content-generic list item.

ListVisuals

Implemented after https://www.figma.com/design/04eHlTWW361rIs3YesfTJo/Data-platform?node-id=813-9806&t=Kofxiju5Tn4DszG2-1

PropertyContent

ListItemContent to display property-like, two-column content.

ShowCollapsingResponse

Responses returned by ListItem::show_hierarchical_with_children.

Enums

DesiredWidth

Traits

ItemButton

A button that can be used by a list item.

ListItemContent

Functions

list_item_scope

Create a scope in which [ListItem]s can be created.