re_viewer::external::re_view
Module controls

Structs

RuntimeModifiers

Constants

ASPECT_SCROLL_MODIFIER

Modifier to press for scroll to change aspect ratio.

DRAG_PAN2D_BUTTON

Which mouse button to drag for panning a 2D view.

DRAG_PAN3D_BUTTON

Which mouse button to drag for panning a 2D view.

HORIZONTAL_SCROLL_MODIFIER

Modifier to press for scroll to pan horizontally.

MOVE_TIME_CURSOR_BUTTON

Clicking this button moves the timeline to where the cursor is.

ROLL_MOUSE

Which mouse button rolls the camera.

ROLL_MOUSE_ALT

Which mouse button rolls the camera if the roll modifier is pressed.

ROLL_MOUSE_MODIFIER

See ROLL_MOUSE_ALT.

ROTATE3D_BUTTON

Which mouse button to drag for rotating a 3D view.

SELECTION_RECT_ZOOM_BUTTON

Rectangles drawn with this mouse button zoom in 2D views.

SPEED_UP_3D_MODIFIER

Which modifier speeds up the 3D camera movement.

TRACKED_OBJECT_RESTORE_KEY

Key to restore the camera.

ZOOM_SCROLL_MODIFIER

Modifier to press for scroll to zoom.