re_viewer::external::re_view::external::re_chunk_store::external::re_chunk
Struct EntityPath

pub struct EntityPath {
    hash: EntityPathHash,
    parts: Arc<Vec<EntityPathPart>>,
}

The unique identifier of an entity, e.g. camera/3/points

The entity path is a list of parts separated by slashes. Each part is a non-empty string, that can contain any character. When written as a string, some characters in the parts need to be escaped with a \ (only character, numbers, ., -, _ does not need escaping).

See https://www.rerun.io/docs/concepts/entity-path for more on entity paths.

This is basically implemented as a list of strings, but is reference-counted internally, so it is cheap to clone. It also has a precomputed hash and implemented nohash_hasher::IsEnabled, so it is very cheap to use in a nohash_hasher::IntMap and nohash_hasher::IntSet.

assert_eq!(
    EntityPath::parse_strict(r#"camera/ACME\ Örnöga/points/42"#).unwrap(),
    EntityPath::new(vec![
        "camera".into(),
        "ACME Örnöga".into(),
        "points".into(),
        "42".into(),
    ])
);

Fields
hash: EntityPathHashparts: Arc<Vec<EntityPathPart>>
Implementations

impl EntityPath

pub fn root() -> EntityPath

pub fn properties() -> EntityPath

The reserved namespace for properties.

pub fn recording_properties() -> EntityPath

The reserved namespace for the RecordingProperties that are specific to the Rerun viewer.

pub fn is_reserved(&self) -> bool

Returns true if the EntityPath belongs to a reserved namespace.

Returns true iff the root entity starts with __.

pub fn new(parts: Vec<EntityPathPart>) -> EntityPath

pub fn from_file_path_as_single_string(file_path: &Path) -> EntityPath

Treat the file path as one opaque string.

The file path separators will NOT become splits in the new path. The returned path will only have one part.

pub fn from_file_path(file_path: &Path) -> EntityPath

Treat the file path as an entity path hierarchy.

The file path separators will become splits in the new path.

pub fn from_single_string(string: impl Into<InternedString>) -> EntityPath

Treat the string as one opaque string, NOT splitting on any slashes.

The given string is expected to be unescaped, i.e. any \ is treated as a normal character.

pub fn iter(&self) -> impl DoubleEndedIterator + ExactSizeIterator

pub fn last(&self) -> Option<&EntityPathPart>

pub fn as_slice(&self) -> &[EntityPathPart]

pub fn to_vec(&self) -> Vec<EntityPathPart>

pub fn is_root(&self) -> bool

pub fn starts_with(&self, prefix: &EntityPath) -> bool

Is this equals to, or a descendant of, the given path.

pub fn is_descendant_of(&self, other: &EntityPath) -> bool

Is this a strict descendant of the given path.

pub fn is_child_of(&self, other: &EntityPath) -> bool

Is this a direct child of the other path.

pub fn len(&self) -> usize

Number of parts

pub fn hash(&self) -> EntityPathHash

pub fn hash64(&self) -> u64

Precomputed 64-bit hash.

pub fn parent(&self) -> Option<EntityPath>

Return None if root.

pub fn join(&self, other: &EntityPath) -> EntityPath

pub fn incremental_walk<'a>( start: Option<&EntityPath>, end: &'a EntityPath, ) -> impl Iterator<Item = EntityPath> + 'a + use<'a>

Helper function to iterate over all incremental EntityPaths from start to end, NOT including start itself.

For example incremental_walk("foo", "foo/bar/baz") returns: ["foo/bar", "foo/bar/baz"]

pub fn common_ancestor(&self, other: &EntityPath) -> EntityPath

Returns the first common ancestor of two paths.

If both paths are the same, the common ancestor is the path itself.

pub fn common_ancestor_of<'a>( entities: impl Iterator<Item = &'a EntityPath>, ) -> EntityPath

Returns the first common ancestor of a list of entity paths.

pub fn short_names_with_disambiguation( entities: impl IntoIterator<Item = EntityPath>, ) -> HashMap<EntityPath, String, RandomState>

Returns short names for a collection of entities based on the last part(s), ensuring uniqueness. Disambiguation is achieved by increasing the number of entity parts used.

Note: the result is undefined when the input contains duplicates.

impl EntityPath

Entity path parsing

When parsing a DataPath, it is important that we can distinguish the component and index from the actual entity path. This requires us to forbid certain characters in an entity part name. For instance, in foo/bar.baz, is baz a component name, or part of the entity path? So, when parsing a full DataPaths we are quite strict with what we allow. But when parsing EntityPaths we want to be a bit more forgiving, so we can accept things like foo/bar.baz and transform it into foo/"bar.baz". This allows user to do things like log(f"foo/{filename}", my_mesh) without Rerun throwing a fit.

pub fn parse_strict(input: &str) -> Result<EntityPath, PathParseError>

Parse an entity path from a string, with strict checks for correctness.

Parses anything that ent_path.to_string() outputs.

For a forgiving parse that accepts anything, use Self::parse_forgiving.

pub fn parse_forgiving(input: &str) -> EntityPath

Parses an entity path, handling any malformed input with a logged warning.

Things like foo/Hallå Där! will be accepted, and transformed into the path foo/Hallå\ Där\!.

For a strict parses, use Self::parse_strict instead.