re_viewer::external::re_view::external::re_query::range_zip
Function range_zip_2x9

pub fn range_zip_2x9<Idx, IR0, R0, IR1, R1, IO0, O0, IO1, O1, IO2, O2, IO3, O3, IO4, O4, IO5, O5, IO6, O6, IO7, O7, IO8, O8>(
    r0: IR0,
    r1: IR1,
    o0: IO0,
    o1: IO1,
    o2: IO2,
    o3: IO3,
    o4: IO4,
    o5: IO5,
    o6: IO6,
    o7: IO7,
    o8: IO8,
) -> RangeZip2x9<Idx, <IR0 as IntoIterator>::IntoIter, R0, <IR1 as IntoIterator>::IntoIter, R1, <IO0 as IntoIterator>::IntoIter, O0, <IO1 as IntoIterator>::IntoIter, O1, <IO2 as IntoIterator>::IntoIter, O2, <IO3 as IntoIterator>::IntoIter, O3, <IO4 as IntoIterator>::IntoIter, O4, <IO5 as IntoIterator>::IntoIter, O5, <IO6 as IntoIterator>::IntoIter, O6, <IO7 as IntoIterator>::IntoIter, O7, <IO8 as IntoIterator>::IntoIter, O8> 
where
    Idx: Ord,
    IR0: IntoIterator<Item = (Idx, R0)>,
    IR1: IntoIterator<Item = (Idx, R1)>,
    IO0: IntoIterator<Item = (Idx, O0)>,
    IO1: IntoIterator<Item = (Idx, O1)>,
    IO2: IntoIterator<Item = (Idx, O2)>,
    IO3: IntoIterator<Item = (Idx, O3)>,
    IO4: IntoIterator<Item = (Idx, O4)>,
    IO5: IntoIterator<Item = (Idx, O5)>,
    IO6: IntoIterator<Item = (Idx, O6)>,
    IO7: IntoIterator<Item = (Idx, O7)>,
    IO8: IntoIterator<Item = (Idx, O8)>,

Returns a new RangeZip2x9 iterator.

The number of elements in a range zip iterator corresponds to the number of elements in the shortest of its required iterators (r0, r1).

Each call to next is guaranteed to yield the next value for each required iterator, as well as the most recent index amongst all of them.

Optional iterators accumulate their state and yield their most recent value (if any), each time the required iterators fire.