re_viewer::external::tokio::sync::mpsc
Function unbounded_channel

pub fn unbounded_channel<T>() -> (UnboundedSender<T>, UnboundedReceiver<T>)

Creates an unbounded mpsc channel for communicating between asynchronous tasks without backpressure.

A send on this channel will always succeed as long as the receive half has not been closed. If the receiver falls behind, messages will be arbitrarily buffered.

Note that the amount of available system memory is an implicit bound to the channel. Using an unbounded channel has the ability of causing the process to run out of memory. In this case, the process will be aborted.