re_viewer::external::tokio::task
Function unconstrained

pub fn unconstrained<F>(inner: F) -> Unconstrained<F> 

Turn off cooperative scheduling for a future. The future will never be forced to yield by Tokio. Using this exposes your service to starvation if the unconstrained future never yields otherwise.

See also the usage example in the task module.