Module re_viewer::math

Structs

BoundingBox

A 3-dimensional axis-aligned bounding box

Rect

A rectangular region of space.

Functions

ease_out

line_segment_distance_sq_to_point_2d

line_segment_distance_sq_to_point_3d

line_segment_distance_to_point_3d

max 🔒

min 🔒

ray_bbox_intersect

Returns the distance the ray traveled of the first intersection or f32::INFINITY on miss.

ray_closest_t_line_segment

Compute the distance between a ray and a line segment.

remap

Linearly remap a value from one range to another, so that when x == from.start() returns to.start() and when x == from.end() returns to.end().

remap_clamp

Like remap, but also clamps the value so that the returned value is always in the to range.