Struct re_viewer::math::BoundingBox

pub struct BoundingBox {
    pub min: Vec3,
    pub max: Vec3,
}

A 3-dimensional axis-aligned bounding box

Fields
min: Vec3

Bounding box minimum (inclusive).

max: Vec3

Bounding box maximum (inclusive).

Implementations

impl BoundingBox

pub const ZERO: BoundingBox = Self{ min: Vec3::ZERO, max: Vec3::ZERO,}

A BoundingBox that only contains Vec3::ZERO.

pub fn nothing() -> BoundingBox

A BoundingBox that contains no points.

This is useful as the seed for bounding bounding boxes.

pub fn everything() -> BoundingBox

A BoundingBox that contains every point.

pub fn from_min_max(min: Vec3, max: Vec3) -> BoundingBox

Create a bounding box from a minimum and maximum position.

pub fn from_min_size(min: Vec3, size: Vec3) -> BoundingBox

pub fn from_center_size(center: Vec3, size: Vec3) -> BoundingBox

Create a bounding box from a center position and a size.

pub fn from_points(points: impl Iterator<Item = Vec3>) -> BoundingBox

Create a bounding box from an iterator of points that the bounding box will cover.

pub fn center(&self) -> Vec3

Returns the center point of the bounding box.

pub fn size(&self) -> Vec3

Returns the 3D axis size of the bounding box.

pub fn half_size(&self) -> Vec3

Returns half the size (similar to a radius).

pub fn volume(&self) -> f32

Only correct for positively sized boxes.

pub fn is_something(&self) -> bool

True if and only if there is at least one point for which bb.contains(point) is true.

Will return true if Self::min == Self::max. The opposite of is_nothing().

pub fn is_nothing(&self) -> bool

True if and only if there is no point for which bb.contains(point) is true.

The opposite of is_something().

pub fn is_point(&self) -> bool

True if this box contains exactly one point.

true if Self::min == Self::max.

pub fn is_finite(&self) -> bool

Returns true if, and only if, all elements are finite.

If any element is either NaN, positive or negative infinity, this will return false.

pub fn is_nan(&self) -> bool

Returns true if any elements are NaN.

pub fn corners(&self) -> [Vec3; 8]

The eight corners of this bounding box.

pub fn bounding_sphere_radius(&self) -> f32

The minimum radius of a sphere, centered at the origin, fully containing the box.

Requires a well-formed box for the result to be valid.

pub fn centered_bounding_sphere_radius(&self) -> f32

The minimum radius of a sphere, centered at the bounding box, fully containing the box.

Requires a well-formed box for the result to be valid.

pub fn edges(&self) -> [[Vec3; 2]; 12]

The twelve edges of this bounding box.

pub fn extend(&mut self, pos: Vec3)

Enlarge the box to include this point.

pub fn union(self, other: BoundingBox) -> BoundingBox

pub fn intersection(self, other: BoundingBox) -> BoundingBox

Returns the smallest volume that is covered by both self and other, or Self::nothing if the boxes are disjoint.

pub fn contains(&self, point: Vec3) -> bool

Returns true if the point is within (or on the edge of) the box.

pub fn expanded(&self, padding: Vec3) -> BoundingBox

Expand with this much padding on each side.

pub fn translated(&self, translation: Vec3) -> BoundingBox

Translate (move) the box by this much.

pub fn rotated_around_origin(&self, q: &Quat) -> BoundingBox

Return a bounding box that contains this box after it has been rotated around Vec3::ZERO.

Note that the rotated bounding box is very likely larger than the original, since it must be large enough to contain the now rotated box.

pub fn transform_iso(&self, m: &IsoTransform) -> BoundingBox

Return a bounding box that contains this box after it has been transformed.

Note that the rotated bounding box is very likely larger than the original, since it must be large enough to contain the now rotated box.

pub fn transform_affine3(&self, m: &Affine3A) -> BoundingBox

Return a bounding box that contains this box after it has been transformed.

Note that the rotated bounding box is very likely larger than the original, since it must be large enough to contain the now rotated box.

pub fn transform_conformal3(&self, m: &Conformal3) -> BoundingBox

Return a bounding box that contains this box after it has been transformed.

Note that the rotated bounding box is very likely larger than the original, since it must be large enough to contain the now rotated box.