Struct re_viewer::misc::time_control::TimeControl

pub struct TimeControl {
    timeline: Timeline,
    states: BTreeMap<Timeline, TimeState>,
    playing: bool,
    following: bool,
    speed: f32,
    looping: Looping,
}

Controls the global view and progress of the time.

Fields
timeline: Timeline

Name of the timeline (e.g. “log_time”).

states: BTreeMap<Timeline, TimeState>playing: bool

If true, we are either in PlayState::Playing or PlayState::Following.

following: bool

If true, we are in “follow” mode (see PlayState::Following). Ignored when [Self.playing] is false.

speed: f32looping: Looping
Implementations

impl TimeControl

pub fn move_time(
&mut self,
egui_ctx: &Context,
times_per_timeline: &TimesPerTimeline
)

Update the current time

pub fn play_state(&self) -> PlayState

pub fn looping(&self) -> Looping

pub fn set_looping(&mut self, looping: Looping)

pub fn set_play_state(
&mut self,
times_per_timeline: &TimesPerTimeline,
play_state: PlayState
)

pub fn pause(&mut self)

pub fn step_time_back(&mut self, times_per_timeline: &TimesPerTimeline)

pub fn step_time_fwd(&mut self, times_per_timeline: &TimesPerTimeline)

pub fn toggle_play_pause(&mut self, times_per_timeline: &TimesPerTimeline)

pub fn speed(&self) -> f32

playback speed

pub fn set_speed(&mut self, speed: f32)

playback speed

pub fn fps(&self) -> Option<f32>

playback fps

pub fn set_fps(&mut self, fps: f32)

playback fps

pub fn select_a_valid_timeline(&mut self, times_per_timeline: &TimesPerTimeline)

Make sure the selected timeline is a valid one

pub fn timeline(&self) -> &Timeline

The currently selected timeline

pub fn time_type(&self) -> TimeType

The time type of the currently selected timeline

pub fn set_timeline(&mut self, timeline: Timeline)

pub fn time(&self) -> Option<TimeReal>

The current time.

pub fn time_int(&self) -> Option<TimeInt>

The current time.

pub fn time_i64(&self) -> Option<i64>

The current time.

pub fn current_query(&self) -> LatestAtQuery

Query for latest value at the currently selected time on the currently selected timeline.

pub fn active_loop_selection(&self) -> Option<TimeRangeF>

The current loop range, iff selection looping is turned on.

pub fn full_range(
&self,
times_per_timeline: &TimesPerTimeline
) -> Option<TimeRange>

The full range of times for the current timeline

pub fn loop_selection(&self) -> Option<TimeRangeF>

The selected slice of time that is called the “loop selection”.

This can still return Some even if looping is currently off.

pub fn set_loop_selection(&mut self, selection: TimeRangeF)

Set the current loop selection without enabling looping.

pub fn remove_loop_selection(&mut self)

Remove the current loop selection.

pub fn is_time_selected(&self, timeline: &Timeline, needle: TimeInt) -> bool

Is the current time in the selection range (if any), or at the current time mark?

pub fn set_timeline_and_time(
&mut self,
timeline: Timeline,
time: impl Into<TimeReal>
)

pub fn set_time(&mut self, time: impl Into<TimeReal>)

pub(crate) fn time_view(&self) -> Option<TimeView>

The range of time we are currently zoomed in on.

pub(crate) fn set_time_view(&mut self, view: TimeView)

The range of time we are currently zoomed in on.

pub fn reset_time_view(&mut self)

The range of time we are currently zoomed in on.

impl TimeControl

pub fn timeline_selector_ui(
&mut self,
times_per_timeline: &TimesPerTimeline,
ui: &mut Ui
)

pub fn fps_ui(&mut self, ui: &mut Ui)

pub fn play_pause_ui(
&mut self,
re_ui: &ReUi,
times_per_timeline: &TimesPerTimeline,
ui: &mut Ui
)

fn play_button_ui(
&mut self,
re_ui: &ReUi,
ui: &mut Ui,
times_per_timeline: &TimesPerTimeline
)

fn follow_button_ui(
&mut self,
re_ui: &ReUi,
ui: &mut Ui,
times_per_timeline: &TimesPerTimeline
)

fn pause_button_ui(&mut self, re_ui: &ReUi, ui: &mut Ui)

fn step_time_button_ui(
&mut self,
re_ui: &ReUi,
ui: &mut Ui,
times_per_timeline: &TimesPerTimeline
)

fn loop_button_ui(&mut self, re_ui: &ReUi, ui: &mut Ui)

pub fn playback_speed_ui(&mut self, ui: &mut Ui)