Struct re_viewer::store_hub::StoreHub

pub struct StoreHub {
    selected_rec_id: Option<StoreId>,
    selected_application_id: Option<ApplicationId>,
    blueprint_by_app_id: HashMap<ApplicationId, StoreId>,
    store_dbs: StoreBundle,
    was_recording_active: bool,
    blueprint_last_save: HashMap<StoreId, StoreGeneration>,
}

Interface for accessing all blueprints and recordings

The StoreHub provides access to the StoreDb instances that are used to store both blueprints and recordings.

Internally, the StoreHub tracks which ApplicationId and recording id (StoreId) are currently selected in the viewer. These can be configured using StoreHub::set_recording_id and StoreHub::set_app_id respectively.

Fields
selected_rec_id: Option<StoreId>selected_application_id: Option<ApplicationId>blueprint_by_app_id: HashMap<ApplicationId, StoreId>store_dbs: StoreBundlewas_recording_active: bool

Was a recording ever activated? Used by the heuristic controlling the welcome screen.

blueprint_last_save: HashMap<StoreId, StoreGeneration>
Implementations

impl StoreHub

pub fn welcome_screen_app_id() -> ApplicationId

App ID used as a marker to display the welcome screen.

pub fn new() -> Self

Create a new StoreHub.

The StoreHub will contain a single empty blueprint associated with the app ID returned by [StoreHub::welcome_screen_app_id]. It should be used as a marker to display the welcome screen.

pub fn read_context(&mut self) -> Option<StoreContext<'_>>

Get a read-only StoreContext from the StoreHub if one is available.

All of the returned references to blueprints and recordings will have a matching ApplicationId.

pub fn was_recording_active(&self) -> bool

Keeps track if a recording was every activated.

This useful for the heuristic controlling the welcome screen.

pub fn set_recording_id(&mut self, recording_id: StoreId)

Change the selected/visible recording id. This will also change the application-id to match the newly selected recording.

pub fn remove_recording_id(&mut self, recording_id: &StoreId)

pub fn set_app_id(&mut self, app_id: ApplicationId)

Change the selected ApplicationId

pub fn selected_application_id(&self) -> Option<&ApplicationId>

pub fn set_blueprint_for_app_id( &mut self, blueprint_id: StoreId, app_id: ApplicationId )

Change which blueprint is active for a given ApplicationId

pub fn clear_blueprint(&mut self)

Clear the current blueprint

pub fn store_db_mut(&mut self, store_id: &StoreId) -> &mut StoreDb

Mutable access to a StoreDb by id

pub fn purge_empty(&mut self)

Remove any empty StoreDbs from the hub

pub fn purge_fraction_of_ram(&mut self, fraction_to_purge: f32)

Call StoreDb::purge_fraction_of_ram on every recording

pub fn current_recording(&self) -> Option<&StoreDb>

Directly access the StoreDb for the selected recording

pub fn contains_recording(&self, id: &StoreId) -> bool

Check whether the StoreHub contains the referenced recording

pub fn gc_and_persist_app_blueprints(&mut self) -> Result<()>

Persist any in-use blueprints to durable storage.

pub fn try_to_load_persisted_blueprint( &mut self, app_id: &ApplicationId ) -> Result<()>

Try to load the persisted blueprint for the given ApplicationId. Note: If no blueprint exists at the expected path, the result is still considered Ok. It is only an Error if a blueprint exists but fails to load.

pub fn stats(&self) -> StoreHubStats

Populate a StoreHubStats based on the selected app.