Struct re_viewer::App

pub struct App {

    build_info: BuildInfo,
    startup_options: StartupOptions,
    ram_limit_warner: RamLimitWarner,
    re_ui: ReUi,
    text_log_rx: Receiver<LogMsg>,
    component_ui_registry: ComponentUiRegistry,
    rx: Receiver<LogMsg>,
    log_dbs: IntMap<RecordingId, LogDb>,
    state: AppState,
    shutdown: Arc<AtomicBool>,
    pending_promises: HashMap<String, Promise<Box<dyn Any + Send>>>,
    toasts: Toasts,
    memory_panel: MemoryPanel,
    memory_panel_open: bool,
    latest_queue_interest: Instant,
    frame_time_history: History<f32>,
    pending_commands: Vec<Command>,
    cmd_palette: CommandPalette,
    analytics: ViewerAnalytics,
    icon_status: AppIconStatus,

}

The Rerun viewer as an [eframe] application.

Fields
build_info: BuildInfostartup_options: StartupOptionsram_limit_warner: RamLimitWarnerre_ui: ReUitext_log_rx: Receiver<LogMsg>

Listens to the local text log stream

component_ui_registry: ComponentUiRegistryrx: Receiver<LogMsg>log_dbs: IntMap<RecordingId, LogDb>

Where the logs are stored.

state: AppState

What is serialized

shutdown: Arc<AtomicBool>

Set to true on Ctrl-C.

pending_promises: HashMap<String, Promise<Box<dyn Any + Send>>>

Pending background tasks, using poll_promise.

toasts: Toasts

Toast notifications.

memory_panel: MemoryPanelmemory_panel_open: boollatest_queue_interest: Instantframe_time_history: History<f32>

Measures how long a frame takes to paint

pending_commands: Vec<Command>

Commands to run at the end of the frame.

cmd_palette: CommandPaletteanalytics: ViewerAnalyticsicon_status: AppIconStatus
Implementations

impl App

pub fn from_receiver(
build_info: BuildInfo,
app_env: &AppEnvironment,
startup_options: StartupOptions,
re_ui: ReUi,
storage: Option<&dyn Storage>,
rx: Receiver<LogMsg>,
shutdown: Arc<AtomicBool>
) -> Self

Create a viewer that receives new log messages over time

pub fn set_profiler(&mut self, profiler: Profiler)

pub fn spawn_threaded_promise<F, T>(
&mut self,
name: impl Into<String>,
f: F
) -> Result<()>where
F: FnOnce() -> T + Send + 'static,
T: Send + 'static,

Creates a promise with the specified name that will run f on a background thread using the poll_promise crate.

Names can only be re-used once the promise with that name has finished running, otherwise an other is returned.

pub fn poll_promise<T: Any>(&mut self, name: impl AsRef<str>) -> Option<T>

Polls the promise with the given name.

Returns Some<T> it it’s ready, or None otherwise.

Panics if T does not match the actual return value of the promise.

pub fn promise_exists(&mut self, name: impl AsRef<str>) -> bool

Returns whether a promise with the given name is currently running.

fn check_keyboard_shortcuts(&mut self, egui_ctx: &Context)

fn loop_selection(&self) -> Option<(Timeline, TimeRangeF)>

fn run_pending_commands(&mut self, egui_ctx: &Context, frame: &mut Frame)

fn run_command(&mut self, cmd: Command, _frame: &mut Frame, egui_ctx: &Context)

fn run_time_control_command(&mut self, command: TimeControlCommand)

fn selected_app_id(&self) -> ApplicationId

fn blueprint_mut(&mut self, egui_ctx: &Context) -> &mut Blueprint

fn memory_panel_ui(
&mut self,
ui: &mut Ui,
gpu_resource_stats: &WgpuResourcePoolStatistics,
store_stats: &DataStoreStats
)

impl App

fn show_text_logs_as_notifications(&mut self)

Show recent text log messages to the user as toast notifications.

fn receive_messages(&mut self, egui_ctx: &Context)

fn cleanup(&mut self)

fn purge_memory_if_needed(&mut self)

fn reset(&mut self, egui_ctx: &Context)

Reset the viewer to how it looked the first time you ran it.

fn log_db_is_nonempty(&self) -> bool

Do we have an open LogDb that is non-empty?

fn log_db(&mut self) -> &mut LogDb

fn show_log_db(&mut self, log_db: LogDb)

fn handle_dropping_files(&mut self, egui_ctx: &Context)