Module re_viewer::ui::auto_layout

Code for automatic layout of space views.

This uses rough heuristics and have a lot of room for improvement.

Some of the heuristics include:

We want similar space views together. Similar can mean:

Same category (2D vs text vs …)

Similar entity path (common prefix)

We also want to pick aspect ratios that fit the data pretty well

Structs

SpaceMakeInfo

Enums

LayoutSplit 🔒

SplitDirection 🔒

Functions

desired_aspect_ratio 🔒

If we need to pick only one aspect ratio for all these spaces, what is a good aspect ratio?

f 🔒

find_group_split_point 🔒

group_by_category 🔒

group_by_path_prefix 🔒

layout_by_category 🔒

Group categories together, i.e. so that 2D stuff is next to 2D stuff, and text logs are next to text logs.

layout_by_path_prefix 🔒

Put spaces with common path prefix close together.

split_groups 🔒

split_spaces_at 🔒

suggest_split_direction 🔒

tree_from_space_views 🔒

tree_from_split 🔒