Module re_viewer::ui::data_ui

The DataUi trait and implementations provide methods for representing data using [egui].

Modules

annotation_context ğŸ”’

component ğŸ”’

component_path ğŸ”’

component_ui_registry ğŸ”’

data ğŸ”’

entity_path ğŸ”’

image ğŸ”’

instance_path ğŸ”’

log_msg ğŸ”’

msg_id ğŸ”’

Enums

UiVerbosity

Controls how mich space we use to show the data in DataUi.

Traits

DataUi ğŸ”’

Types implementing DataUi can draw themselves with a ViewerContext and [egui::Ui].

Functions

format_cell ğŸ”’