Trait re_viewer::ui::data_ui::DataUi

pub(crate) trait DataUi {
    fn data_ui(
        &self,
        ctx: &mut ViewerContext<'_>,
        ui: &mut Ui,
        verbosity: UiVerbosity,
        query: &LatestAtQuery
    );
}

Types implementing DataUi can draw themselves with a ViewerContext and [egui::Ui].

Required Methods

fn data_ui(
&self,
ctx: &mut ViewerContext<'_>,
ui: &mut Ui,
verbosity: UiVerbosity,
query: &LatestAtQuery
)

If you need to lookup something in the data store, use the given query to do so.