Module re_viewer::ui::query_range_ui

Structs

TimelineSpec 🔒

Compute and store various information about a timeline related to how the UI should behave.

Statics

SPAN_TO_START_TIME_OFFSET_THRESHOLD 🔒

Value of the start time over time span ratio above which an explicit offset is handled.

VISIBLE_HISTORY_SUPPORTED_SPACE_VIEWS 🔒

These space views support the Visible History feature.

Functions

current_range_ui 🔒

query_range_ui 🔒

Draws ui for showing and configuring a query range.

query_range_ui_data_result

query_range_ui_space_view

resolved_visible_history_boundary_ui 🔒

round_down 🔒

round_up 🔒

save_visible_time_ranges 🔒

show_visual_time_range 🔒

space_view_with_visible_history 🔒

time_range_base_time 🔒

time_range_editor 🔒

unit_from_span 🔒

visible_history_boundary_combo_label 🔒

visible_history_boundary_ui 🔒

visible_time_range_ui 🔒

Draws ui for a visible time range from a given override path and a resulting query range.