Module re_viewer::ui::time_panel

Modules

data_density_graph ğŸ”’

Show the data density over time for a data stream.

paint_ticks ğŸ”’

time_axis ğŸ”’

time_ranges_ui ğŸ”’

egui uses f32 precision for all screen-space ui coordinates, which makes sense, because that is plenty of precision for things on screen.

time_selection_ui ğŸ”’

Structs

TimePanel ğŸ”’

A panel that shows entity names to the left, time on the top.

Functions

collapsed_time_marker_and_time ğŸ”’

current_time_ui ğŸ”’

help_button ğŸ”’

initialize_time_ranges_ui ğŸ”’

interact_with_streams_rect ğŸ”’

Returns a scroll delta

is_time_safe_to_show ğŸ”’

A user can drag the time slider to between the timeless data and the first real data.

paint_streams_guide_line ğŸ”’

Painted behind the data density graph.

paint_time_cursor

paint_time_ranges_gaps ğŸ”’

Visually separate the different time segments

time_marker_ui ğŸ”’

A vertical line that shows the current time.

top_row_ui ğŸ”’

view_everything ğŸ”’

Find a nice view of everything.