Module re_viewer::ui::time_panel::data_density_graph

Show the data density over time for a data stream.

The data density is the number of data points per unit of time. We collect this into a histogram, blur it, and then paint it.

Structs

DataDensityGraphPainter

Persistent data for painting the data density graph.

DensityGraph 🔒

Constants

DENSITIES_PER_UI_PIXEL 🔒

Higher = slower, but more accurate.

MARGIN_X 🔒

We need some margin because of the blurring.

Functions

data_density_graph_ui

graph_color 🔒

make_brighter 🔒

show_msg_ids_tooltip 🔒

smooth 🔒

Blur the input slightly.