Module re_viewer::ui::view_tensor::ui

Structs

ColorMapping ğŸ”’

How we map values to colors.

TextureSettings ğŸ”’

Scaling, filtering, aspect ratio, etc for the rendered texture.

ViewTensorState

Enums

ColorMap ğŸ”’

TextureScaling ğŸ”’

Should we scale the rendered texture, and if so, how?

Functions

dimension_name ğŸ”’

image_ui ğŸ”’

into_image ğŸ”’

paint_axis_names ğŸ”’

selected_tensor_slice ğŸ”’

selectors_ui ğŸ”’

slice_ui ğŸ”’

tensor_ui ğŸ”’

view_tensor ğŸ”’