Module re_viewer::ui::viewport

The viewport panel.

Contains all space views.

Structs

TabViewer ğŸ”’

Viewport

Describes the layout and contents of the Viewport Panel.

__FieldVisitor ğŸ”’

__Visitor ğŸ”’

Enums

__Field ğŸ”’

Constants

FIELDS ğŸ”’

Statics

ONCE ğŸ”’

__SEEN_MESSAGES ğŸ”’

Functions

blueprint_row_with_buttons ğŸ”’

Show a single button (add_content), justified, and show a visibility button if the row is hovered.

dim_color ğŸ”’

f ğŸ”’

f ğŸ”’

f ğŸ”’

f ğŸ”’

focus_tab ğŸ”’

help_text_ui ğŸ”’

space_view_ui ğŸ”’

visibility_button_ui ğŸ”’

Type Definitions

VisibilitySet ğŸ”’

What views are visible?