Struct re_viewer::ui::viewport::Viewport

pub struct Viewport {
    space_views: HashMap<SpaceViewId, SpaceView>,
    visible: BTreeSet<SpaceViewId>,
    trees: HashMap<BTreeSet<SpaceViewId>, Tree<SpaceViewId>>,
    maximized: Option<SpaceViewId>,
    has_been_user_edited: bool,
    space_view_entity_window: Option<SpaceViewEntityPicker>,
}

Describes the layout and contents of the Viewport Panel.

Fields
space_views: HashMap<SpaceViewId, SpaceView>

Where the space views are stored.

visible: BTreeSet<SpaceViewId>

Which views are visible.

trees: HashMap<BTreeSet<SpaceViewId>, Tree<SpaceViewId>>

The layouts of all the space views.

One for each combination of what views are visible. So if a user toggles the visibility of one SpaceView, we switch which layout we are using. This is somewhat hacky.

maximized: Option<SpaceViewId>

Show one tab as maximized?

has_been_user_edited: bool

Set to true the first time the user messes around with the viewport blueprint. Before this is set we automatically add new spaces to the viewport when they show up in the data.

space_view_entity_window: Option<SpaceViewEntityPicker>
Implementations

impl Viewport

pub fn new(
ctx: &mut ViewerContext<'_>,
spaces_info: &SpaceInfoCollection
) -> Self

Create a default suggested blueprint using some heuristics.

pub(crate) fn space_view(&self, space_view: &SpaceViewId) -> Option<&SpaceView>

pub(crate) fn space_view_mut(
&mut self,
space_view_id: &SpaceViewId
) -> Option<&mut SpaceView>

pub(crate) fn remove(&mut self, space_view_id: &SpaceViewId) -> Option<SpaceView>

pub fn tree_ui(&mut self, ctx: &mut ViewerContext<'_>, ui: &mut Ui)

Show the blueprint panel tree view.

fn default_open_for_group(group: &DataBlueprintGroup) -> bool

If a group or spaceview has a total of this number of elements, show its subtree by default?

fn space_view_entry_ui(
&mut self,
ctx: &mut ViewerContext<'_>,
ui: &mut Ui,
space_view_id: &SpaceViewId
)

fn data_blueprint_tree_ui(
ctx: &mut ViewerContext<'_>,
ui: &mut Ui,
group_handle: DataBlueprintGroupHandle,
space_view: &mut SpaceView,
space_view_visible: bool
)

pub(crate) fn mark_user_interaction(&mut self)

pub(crate) fn add_space_view(&mut self, space_view: SpaceView) -> SpaceViewId

pub fn show_add_remove_entities_window(&mut self, space_view_id: SpaceViewId)

pub fn on_frame_start(
&mut self,
ctx: &mut ViewerContext<'_>,
spaces_info: &SpaceInfoCollection
)

fn should_auto_add_space_view(&self, space_view_candidate: &SpaceView) -> bool

pub fn viewport_ui(&mut self, ui: &mut Ui, ctx: &mut ViewerContext<'_>)

pub fn add_new_spaceview_button_ui(
&mut self,
ctx: &mut ViewerContext<'_>,
ui: &mut Ui,
spaces_info: &SpaceInfoCollection
)

pub fn space_views_containing_entity_path(
&self,
path: &EntityPath
) -> Vec<SpaceViewId>