Module re_viewer::ui::visible_history

Structs

TimelineSpec 🔒

Compute and store various information about a timeline related to how the UI should behave.

Statics

SPAN_TO_START_TIME_OFFSET_THRESHOLD 🔒

Value of the start time over time span ratio above which an explicit offset is handled.

VISIBLE_HISTORY_SUPPORTED_COMPONENT_NAMES 🔒

Entities containing one of these components support the Visible History feature.

VISIBLE_HISTORY_SUPPORTED_SPACE_VIEWS 🔒

These space views support the Visible History feature.

Functions

current_range_ui 🔒

has_visible_history 🔒

resolved_visible_history_boundary_ui 🔒

round_down 🔒

round_up 🔒

time_range_base_time 🔒

unit_from_span 🔒

visible_history_boundary_combo_label 🔒

visible_history_boundary_ui 🔒

visible_history_ui