Module re_viewer::ui::welcome_screen::example_section

Structs

ExampleDesc 🔒

ExampleDescLayout 🔒

Structure to track both an example description and its layout in the grid.

ExampleSection 🔒

ExampleThumbnail 🔒

Enums

LoadError 🔒

Constants

AFTER_HEADER_VSPACE 🔒

CARD_DESCRIPTION_HEIGHT 🔒

CARD_THUMBNAIL_ASPECT_RATIO 🔒

COLUMN_HSPACE 🔒

DESCRIPTION_INNER_MARGIN 🔒

MAX_COLUMN_COUNT 🔒

MAX_COLUMN_WIDTH 🔒

MIN_COLUMN_WIDTH 🔒

ROW_VSPACE 🔒

THUMBNAIL_RADIUS 🔒

TITLE_TO_GRID_VSPACE 🔒

Functions

default_manifest_url 🔒

load_file_size 🔒

Do a HEAD request to get the size of a file.

load_manifest 🔒

open_example_url 🔒

Type Aliases

Manifest 🔒

ManifestJson 🔒

ManifestPromise 🔒