Crate re_viewer

Rerun Viewer GUI.

This crate contains all the GUI code for the Rerun Viewer, including all 2D and 3D visualization code.

Re-exports

pub use event::SelectionChangeItem;

pub use event::ViewerEvent;

pub use event::ViewerEventKind;

pub use native::run_native_app;

pub use native::run_native_viewer_with_messages;

Modules

app 🔒

app_blueprint 🔒

app_state 🔒

background_tasks 🔒

blueprint

Auto-generated blueprint-related types.

env_vars

event

Viewer event definitions.

external

loading 🔒

native

navigation 🔒

saving 🔒

screenshotter 🔒

Screenshotting not implemented on web yet because we haven’t implemented “copy image to clipboard” there.

startup_options 🔒

ui 🔒

viewer_analytics 🔒

Most analytics events collected by the Rerun Viewer are defined in this file.

Structs

App

The Rerun Viewer as an [eframe] application.

AsyncRuntimeHandle

Thin abstraction over the async runtime.

CommandReceiver

Receiver for the CommandSender

CommandSender

Sender that queues up the execution of commands.

MainThreadToken

A token that (almost) proves we are on the main thread.

StartupOptions

Settings set once at startup (e.g. via command-line options) and not serialized.

Enums

AppEnvironment

Where is this App running in? Used for analytics.

SystemCommand

Commands used by internal system components

Traits

SystemCommandSender

Interface for sending SystemCommand messages.

Functions

build_info

Information about this version of the crate.

command_channel

Creates a new command channel.

customize_eframe_and_setup_renderer

Customize eframe and egui to suit the rerun viewer.

reset_viewer_persistence

Reset the viewer state as stored on disk and local storage, keeping only the analytics state.

wake_up_ui_thread_on_each_msg

This wakes up the ui thread each time we receive a new message.

wake_up_ui_thread_on_each_msg_crossbeam

This wakes up the ui thread each time we receive a new message.

wgpu_options 🔒