Struct re_viewer::app::App

pub struct App {

    build_info: BuildInfo,
    startup_options: StartupOptions,
    start_time: Instant,
    ram_limit_warner: RamLimitWarner,
    pub(crate) re_ui: ReUi,
    screenshotter: Screenshotter,
    profiler: Profiler,
    text_log_rx: Receiver<LogMsg>,
    component_ui_registry: ComponentUiRegistry,
    rx: ReceiveSet<LogMsg>,
    pub(crate) state: AppState,
    pub(crate) background_tasks: BackgroundTasks,
    pub(crate) store_hub: Option<StoreHub>,
    toasts: Toasts,
    memory_panel: MemoryPanel,
    memory_panel_open: bool,
    egui_debug_panel_open: bool,
    pub(crate) latest_queue_interest: Instant,
    pub(crate) frame_time_history: History<f32>,
    pub command_sender: CommandSender,
    command_receiver: CommandReceiver,
    cmd_palette: CommandPalette,
    analytics: ViewerAnalytics,
    space_view_class_registry: SpaceViewClassRegistry,

}

The Rerun Viewer as an [eframe] application.

Fields
build_info: BuildInfostartup_options: StartupOptionsstart_time: Instantram_limit_warner: RamLimitWarnerre_ui: ReUiscreenshotter: Screenshotterprofiler: Profilertext_log_rx: Receiver<LogMsg>

Listens to the local text log stream

component_ui_registry: ComponentUiRegistryrx: ReceiveSet<LogMsg>state: AppState

What is serialized

background_tasks: BackgroundTasks

Pending background tasks, e.g. files being saved.

store_hub: Option<StoreHub>

Interface for all recordings and blueprints

toasts: Toasts

Toast notifications.

memory_panel: MemoryPanelmemory_panel_open: boolegui_debug_panel_open: boollatest_queue_interest: Instantframe_time_history: History<f32>

Measures how long a frame takes to paint

command_sender: CommandSender

Commands to run at the end of the frame.

command_receiver: CommandReceivercmd_palette: CommandPaletteanalytics: ViewerAnalyticsspace_view_class_registry: SpaceViewClassRegistry

All known space view types.

Implementations

impl App

pub fn new( build_info: BuildInfo, app_env: &AppEnvironment, startup_options: StartupOptions, re_ui: ReUi, storage: Option<&dyn Storage> ) -> Self

Create a viewer that receives new log messages over time

pub fn set_profiler(&mut self, profiler: Profiler)

pub fn set_examples_manifest_url(&mut self, url: String)

pub fn build_info(&self) -> &BuildInfo

pub fn re_ui(&self) -> &ReUi

pub fn app_options(&self) -> &AppOptions

pub fn app_options_mut(&mut self) -> &mut AppOptions

pub fn is_screenshotting(&self) -> bool

pub fn add_receiver(&mut self, rx: Receiver<LogMsg>)

pub fn msg_receive_set(&self) -> &ReceiveSet<LogMsg>

pub fn add_space_view_class<T: SpaceViewClass + Default + 'static>( &mut self ) -> Result<(), SpaceViewClassRegistryError>

Adds a new space view class to the viewer.

fn check_keyboard_shortcuts(&self, egui_ctx: &Context)

fn run_pending_system_commands( &mut self, store_hub: &mut StoreHub, egui_ctx: &Context )

fn run_pending_ui_commands( &mut self, egui_ctx: &Context, app_blueprint: &AppBlueprint<'_>, store_context: Option<&StoreContext<'_>> )

fn run_system_command( &mut self, cmd: SystemCommand, store_hub: &mut StoreHub, egui_ctx: &Context )

fn run_ui_command( &mut self, egui_ctx: &Context, app_blueprint: &AppBlueprint<'_>, store_context: Option<&StoreContext<'_>>, cmd: UICommand )

fn run_time_control_command( &mut self, store_context: Option<&StoreContext<'_>>, command: TimeControlCommand )

fn memory_panel_ui( &mut self, ui: &mut Ui, gpu_resource_stats: &WgpuResourcePoolStatistics, store_stats: Option<&StoreHubStats> )

fn egui_debug_panel_ui(&mut self, ui: &mut Ui)

fn ui( &mut self, egui_ctx: &Context, frame: &Frame, app_blueprint: &AppBlueprint<'_>, gpu_resource_stats: &WgpuResourcePoolStatistics, store_context: Option<&StoreContext<'_>>, store_stats: Option<&StoreHubStats> )

Top-level ui function.

Shows the viewer ui.

fn show_text_logs_as_notifications(&mut self)

Show recent text log messages to the user as toast notifications.

fn receive_messages(&mut self, store_hub: &mut StoreHub, egui_ctx: &Context)

fn purge_memory_if_needed(&mut self, store_hub: &mut StoreHub)

fn reset_viewer(&mut self, store_hub: &mut StoreHub, egui_ctx: &Context)

Reset the viewer to how it looked the first time you ran it.

pub fn recording_db(&self) -> Option<&EntityDb>

fn handle_dropping_files(&mut self, egui_ctx: &Context)

fn should_fade_in_welcome_screen(&self) -> bool

fn welcome_screen_opacity(&self, egui_ctx: &Context) -> f32

Handle fading in the welcome screen, if we should.

impl App

pub fn rerun_menu_button_ui( &mut self, frame: &Frame, _store_context: Option<&StoreContext<'_>>, ui: &mut Ui )

fn about_rerun_ui(&self, frame: &Frame, ui: &mut Ui)

fn save_buttons_ui(&mut self, ui: &mut Ui, store_ctx: Option<&StoreContext<'_>>)