Module re_viewer::blueprint::validation_gen

Structs

ActiveTab

Component: The active tab in a tabbed container.

AutoLayout

Component: Whether the viewport layout is determined automatically.

AutoSpaceViews

Component: Whether or not space views should be created automatically.

ColumnShare

Component: The layout share of a column in the container.

EntityPropertiesComponent

Component: The configurable set of overridable properties.

GridColumns

Component: How many columns a grid container should have.

IncludedContent

Component: All the contents in the container.

IncludedSpaceView

Component: The unique id of a space view, used to refer to views in containers.

LockRangeDuringZoom

Component: Indicate whether the range should be locked when zooming in on the data.

PanelExpanded

Component: Whether an application panel is expanded or not.

QueryExpression

Component: An individual QueryExpression used to filter a set of EntityPaths.

RootContainer

Component: The container that sits at the root of a viewport.

RowShare

Component: The layout share of a row in the container.

SpaceViewClass

Component: The class of a SpaceView.

SpaceViewMaximized

Component: Whether a space view is maximized.

SpaceViewOrigin

Component: The origin of a SpaceView.

ViewerRecommendationHash

Component: Hash of a viewer recommendation.

Visible

Component: Whether the container, space view, entity or instance is currently visible.

VisibleTimeRange

Component: The range of values on a given timeline that will be included in a view’s query.

VisualBounds2D

Component: Visual bounds in 2D space used for Spatial2DView.

Enums

BackgroundKind

Component: The type of the background in a view.

ContainerKind

Component: The kind of a blueprint container (tabs, grid, …).

Corner2D

Component: One of four 2D corners, typically used to align objects.

Functions

is_valid_blueprint

Because blueprints are both read and written the schema must match what we expect to find or else we will run into all kinds of problems.