re_viewer
Module external

Re-exports

pub use eframe;

pub use egui;

pub use re_chunk;

pub use re_chunk_store;

pub use re_data_ui;

pub use re_entity_db;

pub use re_log;

pub use re_log_types;

pub use re_memory;

pub use re_renderer;

pub use re_types;

pub use re_ui;

pub use re_viewer_context;

pub use re_viewport;

Modules

arrow

A complete, safe, native Rust implementation of Apache Arrow, a cross-language development platform for in-memory data.

crossbeam

Tools for concurrent programming.

egui_kittest

egui_kittest

nohash_hasher

re_byte_size

Calculate the heap-allocated size of values at runtime.

re_chunk

A chunk of Rerun data, encoded using Arrow. Used for logging, transport, storage and compute.

re_chunk_store

The Rerun chunk store, implemented on top of Apache Arrow using the arrow crate.

re_entity_db

This is how we store and index logging data.

re_log_types

The different types that make up the rerun log format.

re_types

The standard Rerun data types, component types, and archetypes.

re_ui

Rerun GUI theme and helpers, built around egui.

re_view

Rerun View utilities