Module re_viewer::external

Re-exports

pub use re_data_ui;

pub use eframe;

pub use egui;

pub use re_arrow_store;

pub use re_data_store;

pub use re_log;

pub use re_log_types;

pub use re_memory;

pub use re_renderer;

pub use re_types;

pub use re_ui;

pub use re_viewer_context;

pub use re_viewport;

Modules

arrow2

Welcome to arrow2’s documentation. Thanks for checking it out!

nohash_hasher

re_arrow_store

The Rerun datastore, implemented on top of Apache Arrow using the arrow2 crate.

re_data_store

This is how we store and index logging data.

re_log_types

The different types that make up the rerun log format.

re_query

Provide query-centric access to the re_arrow_store.

re_space_view

Rerun Space View utilities

re_ui

Rerun GUI theme and helpers, built around egui.