Crate re_viewer::external::re_entity_db

This is how we store and index logging data.

Feature flags

serde — Enable (de)serialization using serde.

Modules

blueprint

entity_db

entity_properties

entity_tree

external

Structs

EntityDb

An in-memory database built from a stream of LogMsges.

EntityPath

The unique identifier of an entity, e.g. camera/3/points

EntityPathPart

The different parts that make up an EntityPath.

EntityProperties

EntityPropertiesComponent

Component: The configurable set of overridable properties.

EntityPropertyMap

Properties for a collection of entities.

EntityTree

A recursive, manually updated re_data_store::StoreSubscriber that maintains the entity hierarchy.

InstancePath

The path to either a specific instance of an entity, or the whole entity.

InstancePathHash

Hashes of the components of an InstancePath.

StoreBundle

Stores many EntityDbs of recordings and blueprints.

TimeHistogramPerTimeline

Number of messages per time per timeline.

TimeInt

A 64-bit number describing either nanoseconds, sequence numbers or fully static data.

Timeline

A time frame/space, e.g. log_time or frame_nr, coupled with the type of time it keeps.

TimesPerTimeline

A StoreSubscriber that keeps track of all unique timestamps on each Timeline.

VersionedInstancePath

A versioned path (i.e. pinned to a specific RowId) to either a specific instance of an entity, or the whole entity.

VersionedInstancePathHash

Hashes of the components of a VersionedInstancePath.

Enums

ColorMapper

Colormap

EditableAutoValue

A value that is either determined automatically by some heuristic, or specified by the user.

Error

The errors that can occur when misusing the data store.

LegendCorner

Where to put the legend?

StoreLoadError

TimeQuery

A query in time.

TimeSeriesAggregator

What kind of aggregation should be performed when the zoom-level on the X axis goes below 1.0?

Type Aliases

Result

TimeCounts

TimeHistogram

Number of messages per time.