Trait re_viewer::external::re_ui::list_item::ItemButton

pub trait ItemButton {
    // Required method
    fn ui(self: Box<Self>, ui: &mut Ui) -> Response;
}

A button that can be used by a list item.

The button is expected to have a size of crate::DesignTokens::small_icon_size.

How/where the button is displayed is up to the crate::list_item::ListItemContent implementation.

Required Methods

fn ui(self: Box<Self>, ui: &mut Ui) -> Response

Draw the button.