re_viewer
Struct App

pub struct App {

    main_thread_token: MainThreadToken,
    build_info: BuildInfo,
    startup_options: StartupOptions,
    start_time: Instant,
    ram_limit_warner: RamLimitWarner,
    pub(crate) egui_ctx: Context,
    screenshotter: Screenshotter,
    profiler: Profiler,
    text_log_rx: Receiver<LogMsg>,
    component_ui_registry: ComponentUiRegistry,
    rx_log: ReceiveSet<LogMsg>,
    rx_table: Mutex<Vec<Receiver<TableMsg>>>,
    pub(crate) state: AppState,
    pub(crate) background_tasks: BackgroundTasks,
    pub(crate) store_hub: Option<StoreHub>,
    pub(crate) notifications: NotificationUi,
    memory_panel: MemoryPanel,
    memory_panel_open: bool,
    egui_debug_panel_open: bool,
    pub(crate) latest_queue_interest: Instant,
    pub(crate) frame_time_history: History<f32>,
    pub command_sender: CommandSender,
    command_receiver: CommandReceiver,
    cmd_palette: CommandPalette,
    analytics: ViewerAnalytics,
    view_class_registry: ViewClassRegistry,
    pub(crate) panel_state_overrides_active: bool,
    pub(crate) panel_state_overrides: PanelStateOverrides,
    reflection: Reflection,
    pub event_dispatcher: Option<ViewerEventDispatcher>,
    async_runtime: AsyncRuntimeHandle,

}

The Rerun Viewer as an [eframe] application.

Fields
main_thread_token: MainThreadTokenbuild_info: BuildInfostartup_options: StartupOptionsstart_time: Instantram_limit_warner: RamLimitWarneregui_ctx: Contextscreenshotter: Screenshotterprofiler: Profilertext_log_rx: Receiver<LogMsg>

Listens to the local text log stream

component_ui_registry: ComponentUiRegistryrx_log: ReceiveSet<LogMsg>rx_table: Mutex<Vec<Receiver<TableMsg>>>state: AppState

What is serialized

background_tasks: BackgroundTasks

Pending background tasks, e.g. files being saved.

store_hub: Option<StoreHub>

Interface for all recordings and blueprints

notifications: NotificationUi

Notification panel.

memory_panel: MemoryPanelmemory_panel_open: boolegui_debug_panel_open: boollatest_queue_interest: Instantframe_time_history: History<f32>

Measures how long a frame takes to paint

command_sender: CommandSender

Commands to run at the end of the frame.

command_receiver: CommandReceivercmd_palette: CommandPaletteanalytics: ViewerAnalyticsview_class_registry: ViewClassRegistry

All known view types.

panel_state_overrides_active: boolpanel_state_overrides: PanelStateOverridesreflection: Reflectionevent_dispatcher: Option<ViewerEventDispatcher>

External interactions with the Viewer host (JS, custom egui app, notebook, etc.).

async_runtime: AsyncRuntimeHandle

The async runtime that should be used for all asynchronous operations.

Using the global tokio runtime should be avoided since:

we donโt have a tokio runtime on web

we want the user to have full control over the runtime, and not expect that a global runtime exists.

Implementations

impl App

pub fn new( main_thread_token: MainThreadToken, build_info: BuildInfo, app_env: &AppEnvironment, startup_options: StartupOptions, creation_context: &CreationContext<'_>, tokio_runtime: AsyncRuntimeHandle, ) -> Self

pub fn with_commands( main_thread_token: MainThreadToken, build_info: BuildInfo, app_env: &AppEnvironment, startup_options: StartupOptions, creation_context: &CreationContext<'_>, tokio_runtime: AsyncRuntimeHandle, command_channel: (CommandSender, CommandReceiver), ) -> Self

Create a viewer that receives new log messages over time

pub fn set_profiler(&mut self, profiler: Profiler)

pub fn set_examples_manifest_url(&mut self, url: String)

pub fn build_info(&self) -> &BuildInfo

pub fn app_options(&self) -> &AppOptions

pub fn app_options_mut(&mut self) -> &mut AppOptions

pub fn is_screenshotting(&self) -> bool

pub fn add_log_receiver(&mut self, rx: Receiver<LogMsg>)

pub fn add_table_receiver(&mut self, rx: Receiver<TableMsg>)

pub fn msg_receive_set(&self) -> &ReceiveSet<LogMsg>

pub fn add_view_class<T: ViewClass + Default + 'static>( &mut self, ) -> Result<(), ViewClassRegistryError>

๐Deprecated since 0.24.0: Use view_class_registry().add_class(..) instead.

Adds a new view class to the viewer.

pub fn view_class_registry(&mut self) -> &mut ViewClassRegistry

Accesses the view class registry which can be used to extend the Viewer.

WARNING: Many parts or the viewer assume that all views & visualizers are registered before the first frame is rendered. Doing so later in the application life cycle may cause unexpected behavior.

fn check_keyboard_shortcuts(&self, egui_ctx: &Context)

fn run_pending_system_commands( &mut self, store_hub: &mut StoreHub, egui_ctx: &Context, )

fn run_pending_ui_commands( &mut self, egui_ctx: &Context, app_blueprint: &AppBlueprint<'_>, store_context: Option<&StoreContext<'_>>, )

fn run_system_command( &mut self, cmd: SystemCommand, store_hub: &mut StoreHub, egui_ctx: &Context, )

fn go_to_uri_fragment(&self, rec_id: StoreId, fragment: Fragment)

fn run_ui_command( &mut self, egui_ctx: &Context, app_blueprint: &AppBlueprint<'_>, store_context: Option<&StoreContext<'_>>, cmd: UICommand, )

fn run_time_control_command( &mut self, store_context: Option<&StoreContext<'_>>, command: TimeControlCommand, )

fn run_copy_time_range_link_command( &mut self, store_context: Option<&StoreContext<'_>>, )

fn memory_panel_ui( &self, ui: &mut Ui, gpu_resource_stats: &WgpuResourcePoolStatistics, store_stats: Option<&StoreHubStats>, )

fn egui_debug_panel_ui(&self, ui: &mut Ui)

fn ui( &mut self, egui_ctx: &Context, frame: &Frame, app_blueprint: &AppBlueprint<'_>, gpu_resource_stats: &WgpuResourcePoolStatistics, store_context: Option<&StoreContext<'_>>, storage_context: &StorageContext<'_>, store_stats: Option<&StoreHubStats>, )

Top-level ui function.

Shows the viewer ui.

fn show_text_logs_as_notifications(&mut self)

Show recent text log messages to the user as toast notifications.

fn receive_messages(&self, store_hub: &mut StoreHub, egui_ctx: &Context)

fn validate_loaded_events(&self, store_events: &[ChunkStoreEvent])

After loading some data; check if the loaded data makes sense.

fn purge_memory_if_needed(&mut self, store_hub: &mut StoreHub)

fn reset_viewer(&mut self, store_hub: &mut StoreHub, egui_ctx: &Context)

Reset the viewer to how it looked the first time you ran it.

pub fn recording_db(&self) -> Option<&EntityDb>

pub fn recording_config_mut( &mut self, store_hub: &StoreHub, rec_id: &StoreId, ) -> Option<&mut RecordingConfig>

fn handle_dropping_files( egui_ctx: &Context, storage_ctx: &StorageContext<'_>, command_sender: &CommandSender, )

fn should_fade_in_welcome_screen(&self) -> bool

fn welcome_screen_opacity(&self, egui_ctx: &Context) -> f32

Handle fading in the welcome screen, if we should.

pub(crate) fn toggle_fullscreen(&self)

fn process_screenshot_result( &mut self, image: &Arc<ColorImage>, user_data: &UserData, )

pub fn add_redap_server(&self, origin: Origin)

pub fn select_redap_entry(&self, uri: &EntryUri)

impl App

pub fn rerun_menu_button_ui( &mut self, render_state: Option<&RenderState>, _store_context: Option<&StoreContext<'_>>, ui: &mut Ui, )

fn rerun_menu_ui( &mut self, ui: &mut Ui, render_state: Option<&RenderState>, _store_context: Option<&StoreContext<'_>>, )

fn about_rerun_ui(&self, ui: &mut Ui, render_state: Option<&RenderState>)

fn save_buttons_ui(&self, ui: &mut Ui, store_ctx: Option<&StoreContext<'_>>)