Module re_viewer::ui::time_panel

Structs

BallScatterer ğŸ”’

Positions circles on a horizontal line with some vertical scattering to avoid overlap.

Segment ğŸ”’

TimePanel ğŸ”’

A panel that shows entity names to the left, time on the top.

TimeRangesUi ğŸ”’

Represents a compressed view of time. It does so by breaking up the timeline in linear segments.

Constants

GAP_EXPANSION_FRACTION ğŸ”’

How much of the gap use up to expand segments visually to either side?

MAX_GAP ğŸ”’

The ideal gap between time segments.

Functions

current_time_ui ğŸ”’

format_duration

Human-readable description of a duration

gap_width ğŸ”’

Sze of the gap between time segments.

help_button ğŸ”’

initial_time_selection ğŸ”’

initialize_time_ranges_ui ğŸ”’

interact_with_streams_rect ğŸ”’

Returns a scroll delta

loop_selection_ui ğŸ”’

paint_range_text ğŸ”’

paint_ticks ğŸ”’

paint_time_cursor

paint_time_range_ticks ğŸ”’

paint_time_ranges_and_ticks ğŸ”’

paint_time_ranges_gaps ğŸ”’

Visually separate the different time segments

show_data_over_time ğŸ”’

show_msg_ids_tooltip ğŸ”’

time_marker_ui ğŸ”’

A vertical line that shows the current time.

top_row_ui ğŸ”’

view_everything ğŸ”’

Find a nice view of everything.